+++ crates/clickhouse-cloud-api/src/models.rs
@@ -128,14 +128,17 @@ macro_rules! discriminated_union {
 
 mod activity;
 mod api_keys;
+mod backups;
 mod byoc;
 mod invitations;
 mod members;
 mod organization_private_endpoints;
 mod organizations;
+mod postgres;
 mod quotas;
 mod rbac;
 mod scim;
+mod services;
 mod shared;
 mod udfs;
 
@@ -143,6 +146,21 @@ pub use activity::{Activity, ActivityActortype, ActivityKeyupdatetype, ActivityT
 pub use api_keys::{
     ApiKey, ApiKeyHashData, ApiKeyPatchRequest, ApiKeyPatchRequestState, ApiKeyPostRequest,
     ApiKeyPostRequestState, ApiKeyPostResponse, ApiKeyState,
+};
+pub use backups::{
+    AwsBackupBucket, AwsBackupBucketBucketprovider, AwsBackupBucketPatchRequestV1,
+    AwsBackupBucketPatchRequestV1Bucketprovider, AwsBackupBucketPostRequestV1,
+    AwsBackupBucketPostRequestV1Bucketprovider, AwsBackupBucketProperties,
+    AwsBackupBucketPropertiesBucketprovider, AzureBackupBucket, AzureBackupBucketBucketprovider,
+    AzureBackupBucketPatchRequestV1, AzureBackupBucketPatchRequestV1Bucketprovider,
+    AzureBackupBucketPostRequestV1, AzureBackupBucketPostRequestV1Bucketprovider,
+    AzureBackupBucketProperties, AzureBackupBucketPropertiesBucketprovider, Backup, BackupBucket,
+    BackupBucketPatchRequest, BackupBucketPostRequest, BackupBucketProperties, BackupConfiguration,
+    BackupConfigurationPatchRequest, BackupStatus, BackupType, GcpBackupBucket,
+    GcpBackupBucketBucketprovider, GcpBackupBucketPatchRequestV1,
+    GcpBackupBucketPatchRequestV1Bucketprovider, GcpBackupBucketPostRequestV1,
+    GcpBackupBucketPostRequestV1Bucketprovider, GcpBackupBucketProperties,
+    GcpBackupBucketPropertiesBucketprovider,
 };
 pub use byoc::{
     ByocConfig, ByocConfigCloudprovider, ByocConfigRegionid, ByocConfigState,
@@ -160,6 +178,19 @@ pub use organization_private_endpoints::{
     OrganizationPrivateEndpointRegion, OrganizationPrivateEndpointsPatch,
 };
 pub use organizations::{Organization, OrganizationPatchRequest};
+pub use postgres::{
+    BasePostgresService, PgBouncerConfig, PgBouncerConfigResponse, PgConfig,
+    PgConfigDefaultTransactionIsolation, PgConfigResponse, PgConfigSslMinProtocolVersion,
+    PgConfigWalCompression, PgCreatedAtProperty, PgHaType, PgIdProperty, PgIsPrimaryProperty,
+    PgNameProperty, PgPassword, PgPitrRestoreTargetProperty, PgProvider, PgRegion, PgSize,
+    PgStateProperty, PgStorageSize, PgTags, PgTagsResponse, PgVersion, PostgresInstanceConfig,
+    PostgresInstanceConfigResponse, PostgresInstanceUpdateConfigResponse, PostgresMetric,
+    PostgresMetricDataPoint, PostgresMetricSeries, PostgresMetrics, PostgresQueryExecution,
+    PostgresService, PostgresServiceListItem, PostgresServicePasswordResource,
+    PostgresServicePatchRequest, PostgresServicePostRequest, PostgresServiceReadReplicaRequest,
+    PostgresServiceRestoreRequest, PostgresServiceSetPassword, PostgresServiceSetState,
+    PostgresServiceSetStateCommand, PostgresSlowQueryPattern, PostgresSlowQueryPatternDetail,
+};
 pub use quotas::{OrganizationQuota, OrganizationQuotaQuotacode, OrganizationQuotaScope};
 pub use rbac::{
     RBACPolicy, RBACPolicyAllowdeny, RBACPolicyCreateRequest, RBACPolicyCreateRequestAllowdeny,
@@ -178,6 +209,30 @@ pub use scim::{
     ScimUserPhoneNumber, ScimUserPhoto, ScimUserPostRequest, ScimUserPutRequest, ScimUserRole,
     ScimX509Certificate,
 };
+pub use services::{
+    AutoscalingMode, CurrentScaling, CurrentScalingEffectiveautoscalingmode,
+    InstancePrivateEndpoint, InstancePrivateEndpointCloudprovider, InstancePrivateEndpointRegion,
+    InstancePrivateEndpointsPatch, InstanceServiceQueryApiEndpointsPostRequest, InstanceTagsPatch,
+    PrivateEndpointConfig, ScalingSchedule, ScalingScheduleBaseConfig, ScalingScheduleEntry,
+    ScalingScheduleEntryRequest, ScalingSchedulePostRequest, ServicPrivateEndpointePostRequest,
+    Service, ServiceClickhouseSetting, ServiceClickhouseSettingSchemaEntry,
+    ServiceClickhouseSettingWarning, ServiceClickhouseSettingsList,
+    ServiceClickhouseSettingsPatchRequest, ServiceClickhouseSettingsPatchResponse,
+    ServiceClickhouseSettingsSchema, ServiceCompliancetype, ServiceEndpoint, ServiceEndpointChange,
+    ServiceEndpointChangeProtocol, ServiceEndpointProtocol, ServicePasswordPatchRequest,
+    ServicePasswordPatchResponse, ServicePatchRequest, ServicePatchRequestReleasechannel,
+    ServicePostRequest, ServicePostRequestCompliancetype, ServicePostRequestProfile,
+    ServicePostRequestProvider, ServicePostRequestRegion, ServicePostRequestReleasechannel,
+    ServicePostRequestTier, ServicePostResponse, ServiceProfile, ServiceProvider,
+    ServiceQueryAPIEndpoint, ServiceRegion, ServiceReleasechannel,
+    ServiceReplicaScalingPatchRequest, ServiceScalingPatchRequest, ServiceScalingPatchResponse,
+    ServiceScalingPatchResponseCompliancetype, ServiceScalingPatchResponseProfile,
+    ServiceScalingPatchResponseProvider, ServiceScalingPatchResponseRegion,
+    ServiceScalingPatchResponseReleasechannel, ServiceScalingPatchResponseState,
+    ServiceScalingPatchResponseTier, ServiceState, ServiceStatePatchRequest,
+    ServiceStatePatchRequestCommand, ServiceTier, UpgradeWindow, UpgradeWindowPutRequest,
+    UsageCost, UsageCostMetrics, UsageCostRecord, UsageCostRecordEntitytype,
+};
 pub use shared::{
     ApiResponse, AssignedRole, AssignedRoleRoletype, IpAccessListEntry, IpAccessListEntryResponse,
     IpAccessListPatch, License, ResourceTagsV1, ResourceTagsV1Response,
@@ -190,598 +245,6 @@ pub use udfs::{
     UdfVersionCreateRequestV1, UdfVersionCreateRequestV1Type, UdfVersionCreateRequestV2,
     UdfVersionCreateRequestV2Type, UdfVersionListResponse,
 };
-
-/// `pgHaType` enum from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum PgHaType {
-    #[serde(rename = "none")]
-    #[default]
-    None,
-    #[serde(rename = "async")]
-    Async,
-    #[serde(rename = "sync")]
-    Sync,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for PgHaType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::None => write!(f, "none"),
-            Self::Async => write!(f, "async"),
-            Self::Sync => write!(f, "sync"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-impl PgHaType {
-    /// Wire values accepted by the API, excluding the catch-all.
-    pub const VALUES: &'static [&'static str] = &["none", "async", "sync"];
-}
-
-/// `pgProvider` enum from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum PgProvider {
-    #[serde(rename = "aws")]
-    #[default]
-    Aws,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for PgProvider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Aws => write!(f, "aws"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-impl PgProvider {
-    /// Wire values accepted by the API, excluding the catch-all.
-    pub const VALUES: &'static [&'static str] = &["aws"];
-}
-
-/// `pgSize` enum from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum PgSize {
-    #[serde(rename = "c6gd.large")]
-    #[default]
-    C6gd_large,
-    #[serde(rename = "c6gd.xlarge")]
-    C6gd_xlarge,
-    #[serde(rename = "c6gd.2xlarge")]
-    C6gd_2xlarge,
-    #[serde(rename = "c6gd.4xlarge")]
-    C6gd_4xlarge,
-    #[serde(rename = "c6gd.8xlarge")]
-    C6gd_8xlarge,
-    #[serde(rename = "c6gd.16xlarge")]
-    C6gd_16xlarge,
-    #[serde(rename = "i7i.large")]
-    I7i_large,
-    #[serde(rename = "i7i.xlarge")]
-    I7i_xlarge,
-    #[serde(rename = "i7i.2xlarge")]
-    I7i_2xlarge,
-    #[serde(rename = "i7i.4xlarge")]
-    I7i_4xlarge,
-    #[serde(rename = "i7i.8xlarge")]
-    I7i_8xlarge,
-    #[serde(rename = "i7i.12xlarge")]
-    I7i_12xlarge,
-    #[serde(rename = "i7i.16xlarge")]
-    I7i_16xlarge,
-    #[serde(rename = "i7i.24xlarge")]
-    I7i_24xlarge,
-    #[serde(rename = "i7ie.large")]
-    I7ie_large,
-    #[serde(rename = "i7ie.xlarge")]
-    I7ie_xlarge,
-    #[serde(rename = "i7ie.2xlarge")]
-    I7ie_2xlarge,
-    #[serde(rename = "i7ie.3xlarge")]
-    I7ie_3xlarge,
-    #[serde(rename = "i7ie.6xlarge")]
-    I7ie_6xlarge,
-    #[serde(rename = "i7ie.12xlarge")]
-    I7ie_12xlarge,
-    #[serde(rename = "i7ie.18xlarge")]
-    I7ie_18xlarge,
-    #[serde(rename = "i7ie.24xlarge")]
-    I7ie_24xlarge,
-    #[serde(rename = "i8g.large")]
-    I8g_large,
-    #[serde(rename = "i8g.xlarge")]
-    I8g_xlarge,
-    #[serde(rename = "i8g.2xlarge")]
-    I8g_2xlarge,
-    #[serde(rename = "i8g.4xlarge")]
-    I8g_4xlarge,
-    #[serde(rename = "i8g.8xlarge")]
-    I8g_8xlarge,
-    #[serde(rename = "i8g.16xlarge")]
-    I8g_16xlarge,
-    #[serde(rename = "i8g.24xlarge")]
-    I8g_24xlarge,
-    #[serde(rename = "i8ge.large")]
-    I8ge_large,
-    #[serde(rename = "i8ge.xlarge")]
-    I8ge_xlarge,
-    #[serde(rename = "i8ge.2xlarge")]
-    I8ge_2xlarge,
-    #[serde(rename = "i8ge.3xlarge")]
-    I8ge_3xlarge,
-    #[serde(rename = "i8ge.6xlarge")]
-    I8ge_6xlarge,
-    #[serde(rename = "i8ge.12xlarge")]
-    I8ge_12xlarge,
-    #[serde(rename = "i8ge.18xlarge")]
-    I8ge_18xlarge,
-    #[serde(rename = "i8ge.24xlarge")]
-    I8ge_24xlarge,
-    #[serde(rename = "m6gd.large")]
-    M6gd_large,
-    #[serde(rename = "m6gd.xlarge")]
-    M6gd_xlarge,
-    #[serde(rename = "m6gd.2xlarge")]
-    M6gd_2xlarge,
-    #[serde(rename = "m6gd.4xlarge")]
-    M6gd_4xlarge,
-    #[serde(rename = "m6gd.8xlarge")]
-    M6gd_8xlarge,
-    #[serde(rename = "m6gd.16xlarge")]
-    M6gd_16xlarge,
-    #[serde(rename = "m6id.large")]
-    M6id_large,
-    #[serde(rename = "m6id.xlarge")]
-    M6id_xlarge,
-    #[serde(rename = "m6id.2xlarge")]
-    M6id_2xlarge,
-    #[serde(rename = "m6id.4xlarge")]
-    M6id_4xlarge,
-    #[serde(rename = "m6id.8xlarge")]
-    M6id_8xlarge,
-    #[serde(rename = "m6id.16xlarge")]
-    M6id_16xlarge,
-    #[serde(rename = "m8gd.large")]
-    M8gd_large,
-    #[serde(rename = "m8gd.xlarge")]
-    M8gd_xlarge,
-    #[serde(rename = "m8gd.2xlarge")]
-    M8gd_2xlarge,
-    #[serde(rename = "m8gd.4xlarge")]
-    M8gd_4xlarge,
-    #[serde(rename = "m8gd.8xlarge")]
-    M8gd_8xlarge,
-    #[serde(rename = "m8gd.16xlarge")]
-    M8gd_16xlarge,
-    #[serde(rename = "r6gd.medium")]
-    R6gd_medium,
-    #[serde(rename = "r6gd.large")]
-    R6gd_large,
-    #[serde(rename = "r6gd.xlarge")]
-    R6gd_xlarge,
-    #[serde(rename = "r6gd.2xlarge")]
-    R6gd_2xlarge,
-    #[serde(rename = "r6gd.4xlarge")]
-    R6gd_4xlarge,
-    #[serde(rename = "r6gd.8xlarge")]
-    R6gd_8xlarge,
-    #[serde(rename = "r6gd.12xlarge")]
-    R6gd_12xlarge,
-    #[serde(rename = "r6gd.16xlarge")]
-    R6gd_16xlarge,
-    #[serde(rename = "r6id.large")]
-    R6id_large,
-    #[serde(rename = "r6id.xlarge")]
-    R6id_xlarge,
-    #[serde(rename = "r6id.2xlarge")]
-    R6id_2xlarge,
-    #[serde(rename = "r6id.4xlarge")]
-    R6id_4xlarge,
-    #[serde(rename = "r6id.8xlarge")]
-    R6id_8xlarge,
-    #[serde(rename = "r6id.12xlarge")]
-    R6id_12xlarge,
-    #[serde(rename = "r6id.16xlarge")]
-    R6id_16xlarge,
-    #[serde(rename = "r6id.24xlarge")]
-    R6id_24xlarge,
-    #[serde(rename = "r6id.32xlarge")]
-    R6id_32xlarge,
-    #[serde(rename = "r8gd.medium")]
-    R8gd_medium,
-    #[serde(rename = "r8gd.large")]
-    R8gd_large,
-    #[serde(rename = "r8gd.xlarge")]
-    R8gd_xlarge,
-    #[serde(rename = "r8gd.2xlarge")]
-    R8gd_2xlarge,
-    #[serde(rename = "r8gd.4xlarge")]
-    R8gd_4xlarge,
-    #[serde(rename = "r8gd.8xlarge")]
-    R8gd_8xlarge,
-    #[serde(rename = "r8gd.12xlarge")]
-    R8gd_12xlarge,
-    #[serde(rename = "r8gd.16xlarge")]
-    R8gd_16xlarge,
-    #[serde(rename = "r8gd.24xlarge")]
-    R8gd_24xlarge,
-    #[serde(rename = "r8gd.48xlarge")]
-    R8gd_48xlarge,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for PgSize {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::C6gd_large => write!(f, "c6gd.large"),
-            Self::C6gd_xlarge => write!(f, "c6gd.xlarge"),
-            Self::C6gd_2xlarge => write!(f, "c6gd.2xlarge"),
-            Self::C6gd_4xlarge => write!(f, "c6gd.4xlarge"),
-            Self::C6gd_8xlarge => write!(f, "c6gd.8xlarge"),
-            Self::C6gd_16xlarge => write!(f, "c6gd.16xlarge"),
-            Self::I7i_large => write!(f, "i7i.large"),
-            Self::I7i_xlarge => write!(f, "i7i.xlarge"),
-            Self::I7i_2xlarge => write!(f, "i7i.2xlarge"),
-            Self::I7i_4xlarge => write!(f, "i7i.4xlarge"),
-            Self::I7i_8xlarge => write!(f, "i7i.8xlarge"),
-            Self::I7i_12xlarge => write!(f, "i7i.12xlarge"),
-            Self::I7i_16xlarge => write!(f, "i7i.16xlarge"),
-            Self::I7i_24xlarge => write!(f, "i7i.24xlarge"),
-            Self::I7ie_large => write!(f, "i7ie.large"),
-            Self::I7ie_xlarge => write!(f, "i7ie.xlarge"),
-            Self::I7ie_2xlarge => write!(f, "i7ie.2xlarge"),
-            Self::I7ie_3xlarge => write!(f, "i7ie.3xlarge"),
-            Self::I7ie_6xlarge => write!(f, "i7ie.6xlarge"),
-            Self::I7ie_12xlarge => write!(f, "i7ie.12xlarge"),
-            Self::I7ie_18xlarge => write!(f, "i7ie.18xlarge"),
-            Self::I7ie_24xlarge => write!(f, "i7ie.24xlarge"),
-            Self::I8g_large => write!(f, "i8g.large"),
-            Self::I8g_xlarge => write!(f, "i8g.xlarge"),
-            Self::I8g_2xlarge => write!(f, "i8g.2xlarge"),
-            Self::I8g_4xlarge => write!(f, "i8g.4xlarge"),
-            Self::I8g_8xlarge => write!(f, "i8g.8xlarge"),
-            Self::I8g_16xlarge => write!(f, "i8g.16xlarge"),
-            Self::I8g_24xlarge => write!(f, "i8g.24xlarge"),
-            Self::I8ge_large => write!(f, "i8ge.large"),
-            Self::I8ge_xlarge => write!(f, "i8ge.xlarge"),
-            Self::I8ge_2xlarge => write!(f, "i8ge.2xlarge"),
-            Self::I8ge_3xlarge => write!(f, "i8ge.3xlarge"),
-            Self::I8ge_6xlarge => write!(f, "i8ge.6xlarge"),
-            Self::I8ge_12xlarge => write!(f, "i8ge.12xlarge"),
-            Self::I8ge_18xlarge => write!(f, "i8ge.18xlarge"),
-            Self::I8ge_24xlarge => write!(f, "i8ge.24xlarge"),
-            Self::M6gd_large => write!(f, "m6gd.large"),
-            Self::M6gd_xlarge => write!(f, "m6gd.xlarge"),
-            Self::M6gd_2xlarge => write!(f, "m6gd.2xlarge"),
-            Self::M6gd_4xlarge => write!(f, "m6gd.4xlarge"),
-            Self::M6gd_8xlarge => write!(f, "m6gd.8xlarge"),
-            Self::M6gd_16xlarge => write!(f, "m6gd.16xlarge"),
-            Self::M6id_large => write!(f, "m6id.large"),
-            Self::M6id_xlarge => write!(f, "m6id.xlarge"),
-            Self::M6id_2xlarge => write!(f, "m6id.2xlarge"),
-            Self::M6id_4xlarge => write!(f, "m6id.4xlarge"),
-            Self::M6id_8xlarge => write!(f, "m6id.8xlarge"),
-            Self::M6id_16xlarge => write!(f, "m6id.16xlarge"),
-            Self::M8gd_large => write!(f, "m8gd.large"),
-            Self::M8gd_xlarge => write!(f, "m8gd.xlarge"),
-            Self::M8gd_2xlarge => write!(f, "m8gd.2xlarge"),
-            Self::M8gd_4xlarge => write!(f, "m8gd.4xlarge"),
-            Self::M8gd_8xlarge => write!(f, "m8gd.8xlarge"),
-            Self::M8gd_16xlarge => write!(f, "m8gd.16xlarge"),
-            Self::R6gd_medium => write!(f, "r6gd.medium"),
-            Self::R6gd_large => write!(f, "r6gd.large"),
-            Self::R6gd_xlarge => write!(f, "r6gd.xlarge"),
-            Self::R6gd_2xlarge => write!(f, "r6gd.2xlarge"),
-            Self::R6gd_4xlarge => write!(f, "r6gd.4xlarge"),
-            Self::R6gd_8xlarge => write!(f, "r6gd.8xlarge"),
-            Self::R6gd_12xlarge => write!(f, "r6gd.12xlarge"),
-            Self::R6gd_16xlarge => write!(f, "r6gd.16xlarge"),
-            Self::R6id_large => write!(f, "r6id.large"),
-            Self::R6id_xlarge => write!(f, "r6id.xlarge"),
-            Self::R6id_2xlarge => write!(f, "r6id.2xlarge"),
-            Self::R6id_4xlarge => write!(f, "r6id.4xlarge"),
-            Self::R6id_8xlarge => write!(f, "r6id.8xlarge"),
-            Self::R6id_12xlarge => write!(f, "r6id.12xlarge"),
-            Self::R6id_16xlarge => write!(f, "r6id.16xlarge"),
-            Self::R6id_24xlarge => write!(f, "r6id.24xlarge"),
-            Self::R6id_32xlarge => write!(f, "r6id.32xlarge"),
-            Self::R8gd_medium => write!(f, "r8gd.medium"),
-            Self::R8gd_large => write!(f, "r8gd.large"),
-            Self::R8gd_xlarge => write!(f, "r8gd.xlarge"),
-            Self::R8gd_2xlarge => write!(f, "r8gd.2xlarge"),
-            Self::R8gd_4xlarge => write!(f, "r8gd.4xlarge"),
-            Self::R8gd_8xlarge => write!(f, "r8gd.8xlarge"),
-            Self::R8gd_12xlarge => write!(f, "r8gd.12xlarge"),
-            Self::R8gd_16xlarge => write!(f, "r8gd.16xlarge"),
-            Self::R8gd_24xlarge => write!(f, "r8gd.24xlarge"),
-            Self::R8gd_48xlarge => write!(f, "r8gd.48xlarge"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// `pgStateProperty` enum from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum PgStateProperty {
-    #[serde(rename = "creating")]
-    #[default]
-    Creating,
-    #[serde(rename = "restarting")]
-    Restarting,
-    #[serde(rename = "running")]
-    Running,
-    #[serde(rename = "replaying_wal")]
-    Replaying_wal,
-    #[serde(rename = "restoring_backup")]
-    Restoring_backup,
-    #[serde(rename = "finalizing_restore")]
-    Finalizing_restore,
-    #[serde(rename = "unavailable")]
-    Unavailable,
-    #[serde(rename = "stopped")]
-    Stopped,
-    #[serde(rename = "deleting")]
-    Deleting,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for PgStateProperty {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Creating => write!(f, "creating"),
-            Self::Restarting => write!(f, "restarting"),
-            Self::Running => write!(f, "running"),
-            Self::Replaying_wal => write!(f, "replaying_wal"),
-            Self::Restoring_backup => write!(f, "restoring_backup"),
-            Self::Finalizing_restore => write!(f, "finalizing_restore"),
-            Self::Unavailable => write!(f, "unavailable"),
-            Self::Stopped => write!(f, "stopped"),
-            Self::Deleting => write!(f, "deleting"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// `pgVersion` enum from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum PgVersion {
-    #[serde(rename = "18")]
-    #[default]
-    _18,
-    #[serde(rename = "17")]
-    _17,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for PgVersion {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::_18 => write!(f, "18"),
-            Self::_17 => write!(f, "17"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-impl PgVersion {
-    /// Wire values accepted by the API, excluding the catch-all.
-    pub const VALUES: &'static [&'static str] = &["18", "17"];
-}
-
-/// Inline enum for `AwsBackupBucket.bucketProvider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum AwsBackupBucketBucketprovider {
-    #[default]
-    AWS,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for AwsBackupBucketBucketprovider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::AWS => write!(f, "AWS"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `AwsBackupBucketPatchRequestV1.bucketProvider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum AwsBackupBucketPatchRequestV1Bucketprovider {
-    #[default]
-    AWS,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for AwsBackupBucketPatchRequestV1Bucketprovider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::AWS => write!(f, "AWS"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `AwsBackupBucketPostRequestV1.bucketProvider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum AwsBackupBucketPostRequestV1Bucketprovider {
-    #[default]
-    AWS,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for AwsBackupBucketPostRequestV1Bucketprovider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::AWS => write!(f, "AWS"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `AwsBackupBucketProperties.bucketProvider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum AwsBackupBucketPropertiesBucketprovider {
-    #[default]
-    AWS,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for AwsBackupBucketPropertiesBucketprovider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::AWS => write!(f, "AWS"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `AzureBackupBucket.bucketProvider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum AzureBackupBucketBucketprovider {
-    #[default]
-    AZURE,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for AzureBackupBucketBucketprovider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::AZURE => write!(f, "AZURE"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `AzureBackupBucketPatchRequestV1.bucketProvider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum AzureBackupBucketPatchRequestV1Bucketprovider {
-    #[default]
-    AZURE,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for AzureBackupBucketPatchRequestV1Bucketprovider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::AZURE => write!(f, "AZURE"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `AzureBackupBucketPostRequestV1.bucketProvider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum AzureBackupBucketPostRequestV1Bucketprovider {
-    #[default]
-    AZURE,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for AzureBackupBucketPostRequestV1Bucketprovider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::AZURE => write!(f, "AZURE"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `AzureBackupBucketProperties.bucketProvider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum AzureBackupBucketPropertiesBucketprovider {
-    #[default]
-    AZURE,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for AzureBackupBucketPropertiesBucketprovider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::AZURE => write!(f, "AZURE"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `Backup.status`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum BackupStatus {
-    #[serde(rename = "done")]
-    #[default]
-    Done,
-    #[serde(rename = "error")]
-    Error,
-    #[serde(rename = "in_progress")]
-    In_progress,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for BackupStatus {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Done => write!(f, "done"),
-            Self::Error => write!(f, "error"),
-            Self::In_progress => write!(f, "in_progress"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `Backup.type`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum BackupType {
-    #[serde(rename = "full")]
-    #[default]
-    Full,
-    #[serde(rename = "incremental")]
-    Incremental,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for BackupType {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Full => write!(f, "full"),
-            Self::Incremental => write!(f, "incremental"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
 
 /// Inline enum for `ClickPipe.state`.
 #[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
@@ -5251,278 +4714,6 @@ impl std::fmt::Display for CreateReversePrivateEndpointType {
     }
 }
 
-/// `autoscalingMode` enum from the ClickHouse Cloud API.
-///
-/// Used by `Service`, `ServicePostRequest`, `ServiceReplicaScalingPatchRequest`,
-/// `ServiceScalingPatchResponse`, `ScalingScheduleBaseConfig`,
-/// `ScalingScheduleEntry`, and `ScalingScheduleEntryRequest`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum AutoscalingMode {
-    #[serde(rename = "vertical")]
-    #[default]
-    Vertical,
-    #[serde(rename = "horizontal")]
-    Horizontal,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for AutoscalingMode {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Vertical => write!(f, "vertical"),
-            Self::Horizontal => write!(f, "horizontal"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-impl AutoscalingMode {
-    /// Wire values accepted by the API, excluding the catch-all.
-    pub const VALUES: &'static [&'static str] = &["vertical", "horizontal"];
-}
-
-/// Inline enum for `CurrentScaling.effectiveAutoscalingMode`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum CurrentScalingEffectiveautoscalingmode {
-    #[serde(rename = "vertical")]
-    #[default]
-    Vertical,
-    #[serde(rename = "horizontal")]
-    Horizontal,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for CurrentScalingEffectiveautoscalingmode {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Vertical => write!(f, "vertical"),
-            Self::Horizontal => write!(f, "horizontal"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `GcpBackupBucket.bucketProvider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum GcpBackupBucketBucketprovider {
-    #[default]
-    GCP,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for GcpBackupBucketBucketprovider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::GCP => write!(f, "GCP"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `GcpBackupBucketPatchRequestV1.bucketProvider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum GcpBackupBucketPatchRequestV1Bucketprovider {
-    #[default]
-    GCP,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for GcpBackupBucketPatchRequestV1Bucketprovider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::GCP => write!(f, "GCP"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `GcpBackupBucketPostRequestV1.bucketProvider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum GcpBackupBucketPostRequestV1Bucketprovider {
-    #[default]
-    GCP,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for GcpBackupBucketPostRequestV1Bucketprovider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::GCP => write!(f, "GCP"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `GcpBackupBucketProperties.bucketProvider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum GcpBackupBucketPropertiesBucketprovider {
-    #[default]
-    GCP,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for GcpBackupBucketPropertiesBucketprovider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::GCP => write!(f, "GCP"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `InstancePrivateEndpoint.cloudProvider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum InstancePrivateEndpointCloudprovider {
-    #[serde(rename = "gcp")]
-    #[default]
-    Gcp,
-    #[serde(rename = "aws")]
-    Aws,
-    #[serde(rename = "azure")]
-    Azure,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for InstancePrivateEndpointCloudprovider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Gcp => write!(f, "gcp"),
-            Self::Aws => write!(f, "aws"),
-            Self::Azure => write!(f, "azure"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `InstancePrivateEndpoint.region`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum InstancePrivateEndpointRegion {
-    #[serde(rename = "ap-northeast-1")]
-    #[default]
-    Ap_northeast_1,
-    #[serde(rename = "ap-northeast-2")]
-    Ap_northeast_2,
-    #[serde(rename = "ap-south-1")]
-    Ap_south_1,
-    #[serde(rename = "ap-southeast-1")]
-    Ap_southeast_1,
-    #[serde(rename = "ap-southeast-2")]
-    Ap_southeast_2,
-    #[serde(rename = "ca-central-1")]
-    Ca_central_1,
-    #[serde(rename = "eu-central-1")]
-    Eu_central_1,
-    #[serde(rename = "eu-west-1")]
-    Eu_west_1,
-    #[serde(rename = "eu-west-2")]
-    Eu_west_2,
-    #[serde(rename = "il-central-1")]
-    Il_central_1,
-    #[serde(rename = "us-east-1")]
-    Us_east_1,
-    #[serde(rename = "us-east-2")]
-    Us_east_2,
-    #[serde(rename = "us-west-2")]
-    Us_west_2,
-    #[serde(rename = "us-east1")]
-    Us_east1,
-    #[serde(rename = "us-central1")]
-    Us_central1,
-    #[serde(rename = "europe-west2")]
-    Europe_west2,
-    #[serde(rename = "europe-west4")]
-    Europe_west4,
-    #[serde(rename = "asia-southeast1")]
-    Asia_southeast1,
-    #[serde(rename = "asia-northeast1")]
-    Asia_northeast1,
-    #[serde(rename = "eastus")]
-    Eastus,
-    #[serde(rename = "eastus2")]
-    Eastus2,
-    #[serde(rename = "westus3")]
-    Westus3,
-    #[serde(rename = "germanywestcentral")]
-    Germanywestcentral,
-    #[serde(rename = "centralus")]
-    Centralus,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for InstancePrivateEndpointRegion {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Ap_northeast_1 => write!(f, "ap-northeast-1"),
-            Self::Ap_northeast_2 => write!(f, "ap-northeast-2"),
-            Self::Ap_south_1 => write!(f, "ap-south-1"),
-            Self::Ap_southeast_1 => write!(f, "ap-southeast-1"),
-            Self::Ap_southeast_2 => write!(f, "ap-southeast-2"),
-            Self::Ca_central_1 => write!(f, "ca-central-1"),
-            Self::Eu_central_1 => write!(f, "eu-central-1"),
-            Self::Eu_west_1 => write!(f, "eu-west-1"),
-            Self::Eu_west_2 => write!(f, "eu-west-2"),
-            Self::Il_central_1 => write!(f, "il-central-1"),
-            Self::Us_east_1 => write!(f, "us-east-1"),
-            Self::Us_east_2 => write!(f, "us-east-2"),
-            Self::Us_west_2 => write!(f, "us-west-2"),
-            Self::Us_east1 => write!(f, "us-east1"),
-            Self::Us_central1 => write!(f, "us-central1"),
-            Self::Europe_west2 => write!(f, "europe-west2"),
-            Self::Europe_west4 => write!(f, "europe-west4"),
-            Self::Asia_southeast1 => write!(f, "asia-southeast1"),
-            Self::Asia_northeast1 => write!(f, "asia-northeast1"),
-            Self::Eastus => write!(f, "eastus"),
-            Self::Eastus2 => write!(f, "eastus2"),
-            Self::Westus3 => write!(f, "westus3"),
-            Self::Germanywestcentral => write!(f, "germanywestcentral"),
-            Self::Centralus => write!(f, "centralus"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `PostgresServiceSetState.command`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum PostgresServiceSetStateCommand {
-    #[serde(rename = "restart")]
-    #[default]
-    Restart,
-    #[serde(rename = "promote")]
-    Promote,
-    #[serde(rename = "switchover")]
-    Switchover,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for PostgresServiceSetStateCommand {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Restart => write!(f, "restart"),
-            Self::Promote => write!(f, "promote"),
-            Self::Switchover => write!(f, "switchover"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
 /// Inline enum for `ReversePrivateEndpoint.mskAuthentication`.
 #[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
 pub enum ReversePrivateEndpointMskauthentication {
@@ -5597,1323 +4788,6 @@ impl std::fmt::Display for ReversePrivateEndpointType {
             Self::VPC_RESOURCE => write!(f, "VPC_RESOURCE"),
             Self::MSK_MULTI_VPC => write!(f, "MSK_MULTI_VPC"),
             Self::GCP_PSC_SERVICE_ATTACHMENT => write!(f, "GCP_PSC_SERVICE_ATTACHMENT"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `Service.complianceType`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceCompliancetype {
-    #[serde(rename = "hipaa")]
-    #[default]
-    Hipaa,
-    #[serde(rename = "pci")]
-    Pci,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceCompliancetype {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Hipaa => write!(f, "hipaa"),
-            Self::Pci => write!(f, "pci"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `Service.profile`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceProfile {
-    #[serde(rename = "v1-default")]
-    #[default]
-    V1_default,
-    #[serde(rename = "v1-highmem-xs")]
-    V1_highmem_xs,
-    #[serde(rename = "v1-highmem-s")]
-    V1_highmem_s,
-    #[serde(rename = "v1-highmem-m")]
-    V1_highmem_m,
-    #[serde(rename = "v1-highmem-l")]
-    V1_highmem_l,
-    #[serde(rename = "v1-highmem-xl")]
-    V1_highmem_xl,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceProfile {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::V1_default => write!(f, "v1-default"),
-            Self::V1_highmem_xs => write!(f, "v1-highmem-xs"),
-            Self::V1_highmem_s => write!(f, "v1-highmem-s"),
-            Self::V1_highmem_m => write!(f, "v1-highmem-m"),
-            Self::V1_highmem_l => write!(f, "v1-highmem-l"),
-            Self::V1_highmem_xl => write!(f, "v1-highmem-xl"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `Service.provider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceProvider {
-    #[serde(rename = "aws")]
-    #[default]
-    Aws,
-    #[serde(rename = "gcp")]
-    Gcp,
-    #[serde(rename = "azure")]
-    Azure,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceProvider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Aws => write!(f, "aws"),
-            Self::Gcp => write!(f, "gcp"),
-            Self::Azure => write!(f, "azure"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `Service.region`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceRegion {
-    #[serde(rename = "ap-northeast-1")]
-    #[default]
-    Ap_northeast_1,
-    #[serde(rename = "ap-northeast-2")]
-    Ap_northeast_2,
-    #[serde(rename = "ap-south-1")]
-    Ap_south_1,
-    #[serde(rename = "ap-southeast-1")]
-    Ap_southeast_1,
-    #[serde(rename = "ap-southeast-2")]
-    Ap_southeast_2,
-    #[serde(rename = "ca-central-1")]
-    Ca_central_1,
-    #[serde(rename = "eu-central-1")]
-    Eu_central_1,
-    #[serde(rename = "eu-west-1")]
-    Eu_west_1,
-    #[serde(rename = "eu-west-2")]
-    Eu_west_2,
-    #[serde(rename = "il-central-1")]
-    Il_central_1,
-    #[serde(rename = "us-east-1")]
-    Us_east_1,
-    #[serde(rename = "us-east-2")]
-    Us_east_2,
-    #[serde(rename = "us-west-2")]
-    Us_west_2,
-    #[serde(rename = "us-east1")]
-    Us_east1,
-    #[serde(rename = "us-central1")]
-    Us_central1,
-    #[serde(rename = "europe-west2")]
-    Europe_west2,
-    #[serde(rename = "europe-west4")]
-    Europe_west4,
-    #[serde(rename = "asia-southeast1")]
-    Asia_southeast1,
-    #[serde(rename = "asia-northeast1")]
-    Asia_northeast1,
-    #[serde(rename = "eastus")]
-    Eastus,
-    #[serde(rename = "eastus2")]
-    Eastus2,
-    #[serde(rename = "westus3")]
-    Westus3,
-    #[serde(rename = "germanywestcentral")]
-    Germanywestcentral,
-    #[serde(rename = "centralus")]
-    Centralus,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceRegion {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Ap_northeast_1 => write!(f, "ap-northeast-1"),
-            Self::Ap_northeast_2 => write!(f, "ap-northeast-2"),
-            Self::Ap_south_1 => write!(f, "ap-south-1"),
-            Self::Ap_southeast_1 => write!(f, "ap-southeast-1"),
-            Self::Ap_southeast_2 => write!(f, "ap-southeast-2"),
-            Self::Ca_central_1 => write!(f, "ca-central-1"),
-            Self::Eu_central_1 => write!(f, "eu-central-1"),
-            Self::Eu_west_1 => write!(f, "eu-west-1"),
-            Self::Eu_west_2 => write!(f, "eu-west-2"),
-            Self::Il_central_1 => write!(f, "il-central-1"),
-            Self::Us_east_1 => write!(f, "us-east-1"),
-            Self::Us_east_2 => write!(f, "us-east-2"),
-            Self::Us_west_2 => write!(f, "us-west-2"),
-            Self::Us_east1 => write!(f, "us-east1"),
-            Self::Us_central1 => write!(f, "us-central1"),
-            Self::Europe_west2 => write!(f, "europe-west2"),
-            Self::Europe_west4 => write!(f, "europe-west4"),
-            Self::Asia_southeast1 => write!(f, "asia-southeast1"),
-            Self::Asia_northeast1 => write!(f, "asia-northeast1"),
-            Self::Eastus => write!(f, "eastus"),
-            Self::Eastus2 => write!(f, "eastus2"),
-            Self::Westus3 => write!(f, "westus3"),
-            Self::Germanywestcentral => write!(f, "germanywestcentral"),
-            Self::Centralus => write!(f, "centralus"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `Service.releaseChannel`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceReleasechannel {
-    #[serde(rename = "slow")]
-    #[default]
-    Slow,
-    #[serde(rename = "default")]
-    Default,
-    #[serde(rename = "fast")]
-    Fast,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceReleasechannel {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Slow => write!(f, "slow"),
-            Self::Default => write!(f, "default"),
-            Self::Fast => write!(f, "fast"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `Service.state`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceState {
-    #[serde(rename = "starting")]
-    #[default]
-    Starting,
-    #[serde(rename = "stopping")]
-    Stopping,
-    #[serde(rename = "terminating")]
-    Terminating,
-    #[serde(rename = "softdeleting")]
-    Softdeleting,
-    #[serde(rename = "awaking")]
-    Awaking,
-    #[serde(rename = "partially_running")]
-    Partially_running,
-    #[serde(rename = "provisioning")]
-    Provisioning,
-    #[serde(rename = "running")]
-    Running,
-    #[serde(rename = "stopped")]
-    Stopped,
-    #[serde(rename = "terminated")]
-    Terminated,
-    #[serde(rename = "softdeleted")]
-    Softdeleted,
-    #[serde(rename = "degraded")]
-    Degraded,
-    #[serde(rename = "failed")]
-    Failed,
-    #[serde(rename = "idle")]
-    Idle,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceState {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Starting => write!(f, "starting"),
-            Self::Stopping => write!(f, "stopping"),
-            Self::Terminating => write!(f, "terminating"),
-            Self::Softdeleting => write!(f, "softdeleting"),
-            Self::Awaking => write!(f, "awaking"),
-            Self::Partially_running => write!(f, "partially_running"),
-            Self::Provisioning => write!(f, "provisioning"),
-            Self::Running => write!(f, "running"),
-            Self::Stopped => write!(f, "stopped"),
-            Self::Terminated => write!(f, "terminated"),
-            Self::Softdeleted => write!(f, "softdeleted"),
-            Self::Degraded => write!(f, "degraded"),
-            Self::Failed => write!(f, "failed"),
-            Self::Idle => write!(f, "idle"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `Service.tier`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceTier {
-    #[serde(rename = "development")]
-    #[default]
-    Development,
-    #[serde(rename = "production")]
-    Production,
-    #[serde(rename = "dedicated_high_mem")]
-    Dedicated_high_mem,
-    #[serde(rename = "dedicated_high_cpu")]
-    Dedicated_high_cpu,
-    #[serde(rename = "dedicated_standard")]
-    Dedicated_standard,
-    #[serde(rename = "dedicated_standard_n2d_standard_4")]
-    Dedicated_standard_n2d_standard_4,
-    #[serde(rename = "dedicated_standard_n2d_standard_8")]
-    Dedicated_standard_n2d_standard_8,
-    #[serde(rename = "dedicated_standard_n2d_standard_32")]
-    Dedicated_standard_n2d_standard_32,
-    #[serde(rename = "dedicated_standard_n2d_standard_128")]
-    Dedicated_standard_n2d_standard_128,
-    #[serde(rename = "dedicated_standard_n2d_standard_32_16SSD")]
-    Dedicated_standard_n2d_standard_32_16SSD,
-    #[serde(rename = "dedicated_standard_n2d_standard_64_24SSD")]
-    Dedicated_standard_n2d_standard_64_24SSD,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceTier {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Development => write!(f, "development"),
-            Self::Production => write!(f, "production"),
-            Self::Dedicated_high_mem => write!(f, "dedicated_high_mem"),
-            Self::Dedicated_high_cpu => write!(f, "dedicated_high_cpu"),
-            Self::Dedicated_standard => write!(f, "dedicated_standard"),
-            Self::Dedicated_standard_n2d_standard_4 => {
-                write!(f, "dedicated_standard_n2d_standard_4")
-            }
-            Self::Dedicated_standard_n2d_standard_8 => {
-                write!(f, "dedicated_standard_n2d_standard_8")
-            }
-            Self::Dedicated_standard_n2d_standard_32 => {
-                write!(f, "dedicated_standard_n2d_standard_32")
-            }
-            Self::Dedicated_standard_n2d_standard_128 => {
-                write!(f, "dedicated_standard_n2d_standard_128")
-            }
-            Self::Dedicated_standard_n2d_standard_32_16SSD => {
-                write!(f, "dedicated_standard_n2d_standard_32_16SSD")
-            }
-            Self::Dedicated_standard_n2d_standard_64_24SSD => {
-                write!(f, "dedicated_standard_n2d_standard_64_24SSD")
-            }
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ServiceEndpoint.protocol`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceEndpointProtocol {
-    #[serde(rename = "https")]
-    #[default]
-    Https,
-    #[serde(rename = "nativesecure")]
-    Nativesecure,
-    #[serde(rename = "mysql")]
-    Mysql,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceEndpointProtocol {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Https => write!(f, "https"),
-            Self::Nativesecure => write!(f, "nativesecure"),
-            Self::Mysql => write!(f, "mysql"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ServiceEndpointChange.protocol`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceEndpointChangeProtocol {
-    #[serde(rename = "mysql")]
-    #[default]
-    Mysql,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceEndpointChangeProtocol {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Mysql => write!(f, "mysql"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ServicePatchRequest.releaseChannel`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServicePatchRequestReleasechannel {
-    #[serde(rename = "slow")]
-    #[default]
-    Slow,
-    #[serde(rename = "default")]
-    Default,
-    #[serde(rename = "fast")]
-    Fast,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServicePatchRequestReleasechannel {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Slow => write!(f, "slow"),
-            Self::Default => write!(f, "default"),
-            Self::Fast => write!(f, "fast"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-impl ServicePatchRequestReleasechannel {
-    /// Wire values accepted by the API, excluding the catch-all.
-    pub const VALUES: &'static [&'static str] = &["slow", "default", "fast"];
-}
-
-/// Inline enum for `ServicePostRequest.complianceType`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServicePostRequestCompliancetype {
-    #[serde(rename = "hipaa")]
-    #[default]
-    Hipaa,
-    #[serde(rename = "pci")]
-    Pci,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServicePostRequestCompliancetype {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Hipaa => write!(f, "hipaa"),
-            Self::Pci => write!(f, "pci"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-impl ServicePostRequestCompliancetype {
-    /// Wire values accepted by the API, excluding the catch-all.
-    pub const VALUES: &'static [&'static str] = &["hipaa", "pci"];
-}
-
-/// Inline enum for `ServicePostRequest.profile`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServicePostRequestProfile {
-    #[serde(rename = "v1-default")]
-    #[default]
-    V1_default,
-    #[serde(rename = "v1-highmem-xs")]
-    V1_highmem_xs,
-    #[serde(rename = "v1-highmem-s")]
-    V1_highmem_s,
-    #[serde(rename = "v1-highmem-m")]
-    V1_highmem_m,
-    #[serde(rename = "v1-highmem-l")]
-    V1_highmem_l,
-    #[serde(rename = "v1-highmem-xl")]
-    V1_highmem_xl,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServicePostRequestProfile {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::V1_default => write!(f, "v1-default"),
-            Self::V1_highmem_xs => write!(f, "v1-highmem-xs"),
-            Self::V1_highmem_s => write!(f, "v1-highmem-s"),
-            Self::V1_highmem_m => write!(f, "v1-highmem-m"),
-            Self::V1_highmem_l => write!(f, "v1-highmem-l"),
-            Self::V1_highmem_xl => write!(f, "v1-highmem-xl"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-impl ServicePostRequestProfile {
-    /// Wire values accepted by the API, excluding the catch-all.
-    pub const VALUES: &'static [&'static str] = &[
-        "v1-default",
-        "v1-highmem-xs",
-        "v1-highmem-s",
-        "v1-highmem-m",
-        "v1-highmem-l",
-        "v1-highmem-xl",
-    ];
-}
-
-/// Inline enum for `ServicePostRequest.provider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServicePostRequestProvider {
-    #[serde(rename = "aws")]
-    #[default]
-    Aws,
-    #[serde(rename = "gcp")]
-    Gcp,
-    #[serde(rename = "azure")]
-    Azure,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServicePostRequestProvider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Aws => write!(f, "aws"),
-            Self::Gcp => write!(f, "gcp"),
-            Self::Azure => write!(f, "azure"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-impl ServicePostRequestProvider {
-    /// Wire values accepted by the API, excluding the catch-all.
-    pub const VALUES: &'static [&'static str] = &["aws", "gcp", "azure"];
-}
-
-/// Inline enum for `ServicePostRequest.region`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServicePostRequestRegion {
-    #[serde(rename = "ap-northeast-1")]
-    #[default]
-    Ap_northeast_1,
-    #[serde(rename = "ap-northeast-2")]
-    Ap_northeast_2,
-    #[serde(rename = "ap-south-1")]
-    Ap_south_1,
-    #[serde(rename = "ap-southeast-1")]
-    Ap_southeast_1,
-    #[serde(rename = "ap-southeast-2")]
-    Ap_southeast_2,
-    #[serde(rename = "ca-central-1")]
-    Ca_central_1,
-    #[serde(rename = "eu-central-1")]
-    Eu_central_1,
-    #[serde(rename = "eu-west-1")]
-    Eu_west_1,
-    #[serde(rename = "eu-west-2")]
-    Eu_west_2,
-    #[serde(rename = "il-central-1")]
-    Il_central_1,
-    #[serde(rename = "us-east-1")]
-    Us_east_1,
-    #[serde(rename = "us-east-2")]
-    Us_east_2,
-    #[serde(rename = "us-west-2")]
-    Us_west_2,
-    #[serde(rename = "us-east1")]
-    Us_east1,
-    #[serde(rename = "us-central1")]
-    Us_central1,
-    #[serde(rename = "europe-west2")]
-    Europe_west2,
-    #[serde(rename = "europe-west4")]
-    Europe_west4,
-    #[serde(rename = "asia-southeast1")]
-    Asia_southeast1,
-    #[serde(rename = "asia-northeast1")]
-    Asia_northeast1,
-    #[serde(rename = "eastus")]
-    Eastus,
-    #[serde(rename = "eastus2")]
-    Eastus2,
-    #[serde(rename = "westus3")]
-    Westus3,
-    #[serde(rename = "germanywestcentral")]
-    Germanywestcentral,
-    #[serde(rename = "centralus")]
-    Centralus,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServicePostRequestRegion {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Ap_northeast_1 => write!(f, "ap-northeast-1"),
-            Self::Ap_northeast_2 => write!(f, "ap-northeast-2"),
-            Self::Ap_south_1 => write!(f, "ap-south-1"),
-            Self::Ap_southeast_1 => write!(f, "ap-southeast-1"),
-            Self::Ap_southeast_2 => write!(f, "ap-southeast-2"),
-            Self::Ca_central_1 => write!(f, "ca-central-1"),
-            Self::Eu_central_1 => write!(f, "eu-central-1"),
-            Self::Eu_west_1 => write!(f, "eu-west-1"),
-            Self::Eu_west_2 => write!(f, "eu-west-2"),
-            Self::Il_central_1 => write!(f, "il-central-1"),
-            Self::Us_east_1 => write!(f, "us-east-1"),
-            Self::Us_east_2 => write!(f, "us-east-2"),
-            Self::Us_west_2 => write!(f, "us-west-2"),
-            Self::Us_east1 => write!(f, "us-east1"),
-            Self::Us_central1 => write!(f, "us-central1"),
-            Self::Europe_west2 => write!(f, "europe-west2"),
-            Self::Europe_west4 => write!(f, "europe-west4"),
-            Self::Asia_southeast1 => write!(f, "asia-southeast1"),
-            Self::Asia_northeast1 => write!(f, "asia-northeast1"),
-            Self::Eastus => write!(f, "eastus"),
-            Self::Eastus2 => write!(f, "eastus2"),
-            Self::Westus3 => write!(f, "westus3"),
-            Self::Germanywestcentral => write!(f, "germanywestcentral"),
-            Self::Centralus => write!(f, "centralus"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-impl ServicePostRequestRegion {
-    /// Wire values accepted by the API, excluding the catch-all.
-    pub const VALUES: &'static [&'static str] = &[
-        "ap-northeast-1",
-        "ap-northeast-2",
-        "ap-south-1",
-        "ap-southeast-1",
-        "ap-southeast-2",
-        "ca-central-1",
-        "eu-central-1",
-        "eu-west-1",
-        "eu-west-2",
-        "il-central-1",
-        "us-east-1",
-        "us-east-2",
-        "us-west-2",
-        "us-east1",
-        "us-central1",
-        "europe-west2",
-        "europe-west4",
-        "asia-southeast1",
-        "asia-northeast1",
-        "eastus",
-        "eastus2",
-        "westus3",
-        "germanywestcentral",
-        "centralus",
-    ];
-}
-
-/// Inline enum for `ServicePostRequest.releaseChannel`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServicePostRequestReleasechannel {
-    #[serde(rename = "slow")]
-    #[default]
-    Slow,
-    #[serde(rename = "default")]
-    Default,
-    #[serde(rename = "fast")]
-    Fast,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServicePostRequestReleasechannel {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Slow => write!(f, "slow"),
-            Self::Default => write!(f, "default"),
-            Self::Fast => write!(f, "fast"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-impl ServicePostRequestReleasechannel {
-    /// Wire values accepted by the API, excluding the catch-all.
-    pub const VALUES: &'static [&'static str] = &["slow", "default", "fast"];
-}
-
-/// Inline enum for `ServicePostRequest.tier`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServicePostRequestTier {
-    #[serde(rename = "development")]
-    #[default]
-    Development,
-    #[serde(rename = "production")]
-    Production,
-    #[serde(rename = "dedicated_high_mem")]
-    Dedicated_high_mem,
-    #[serde(rename = "dedicated_high_cpu")]
-    Dedicated_high_cpu,
-    #[serde(rename = "dedicated_standard")]
-    Dedicated_standard,
-    #[serde(rename = "dedicated_standard_n2d_standard_4")]
-    Dedicated_standard_n2d_standard_4,
-    #[serde(rename = "dedicated_standard_n2d_standard_8")]
-    Dedicated_standard_n2d_standard_8,
-    #[serde(rename = "dedicated_standard_n2d_standard_32")]
-    Dedicated_standard_n2d_standard_32,
-    #[serde(rename = "dedicated_standard_n2d_standard_128")]
-    Dedicated_standard_n2d_standard_128,
-    #[serde(rename = "dedicated_standard_n2d_standard_32_16SSD")]
-    Dedicated_standard_n2d_standard_32_16SSD,
-    #[serde(rename = "dedicated_standard_n2d_standard_64_24SSD")]
-    Dedicated_standard_n2d_standard_64_24SSD,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServicePostRequestTier {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Development => write!(f, "development"),
-            Self::Production => write!(f, "production"),
-            Self::Dedicated_high_mem => write!(f, "dedicated_high_mem"),
-            Self::Dedicated_high_cpu => write!(f, "dedicated_high_cpu"),
-            Self::Dedicated_standard => write!(f, "dedicated_standard"),
-            Self::Dedicated_standard_n2d_standard_4 => {
-                write!(f, "dedicated_standard_n2d_standard_4")
-            }
-            Self::Dedicated_standard_n2d_standard_8 => {
-                write!(f, "dedicated_standard_n2d_standard_8")
-            }
-            Self::Dedicated_standard_n2d_standard_32 => {
-                write!(f, "dedicated_standard_n2d_standard_32")
-            }
-            Self::Dedicated_standard_n2d_standard_128 => {
-                write!(f, "dedicated_standard_n2d_standard_128")
-            }
-            Self::Dedicated_standard_n2d_standard_32_16SSD => {
-                write!(f, "dedicated_standard_n2d_standard_32_16SSD")
-            }
-            Self::Dedicated_standard_n2d_standard_64_24SSD => {
-                write!(f, "dedicated_standard_n2d_standard_64_24SSD")
-            }
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ServiceScalingPatchResponse.complianceType`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceScalingPatchResponseCompliancetype {
-    #[serde(rename = "hipaa")]
-    #[default]
-    Hipaa,
-    #[serde(rename = "pci")]
-    Pci,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceScalingPatchResponseCompliancetype {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Hipaa => write!(f, "hipaa"),
-            Self::Pci => write!(f, "pci"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ServiceScalingPatchResponse.profile`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceScalingPatchResponseProfile {
-    #[serde(rename = "v1-default")]
-    #[default]
-    V1_default,
-    #[serde(rename = "v1-highmem-xs")]
-    V1_highmem_xs,
-    #[serde(rename = "v1-highmem-s")]
-    V1_highmem_s,
-    #[serde(rename = "v1-highmem-m")]
-    V1_highmem_m,
-    #[serde(rename = "v1-highmem-l")]
-    V1_highmem_l,
-    #[serde(rename = "v1-highmem-xl")]
-    V1_highmem_xl,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceScalingPatchResponseProfile {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::V1_default => write!(f, "v1-default"),
-            Self::V1_highmem_xs => write!(f, "v1-highmem-xs"),
-            Self::V1_highmem_s => write!(f, "v1-highmem-s"),
-            Self::V1_highmem_m => write!(f, "v1-highmem-m"),
-            Self::V1_highmem_l => write!(f, "v1-highmem-l"),
-            Self::V1_highmem_xl => write!(f, "v1-highmem-xl"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ServiceScalingPatchResponse.provider`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceScalingPatchResponseProvider {
-    #[serde(rename = "aws")]
-    #[default]
-    Aws,
-    #[serde(rename = "gcp")]
-    Gcp,
-    #[serde(rename = "azure")]
-    Azure,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceScalingPatchResponseProvider {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Aws => write!(f, "aws"),
-            Self::Gcp => write!(f, "gcp"),
-            Self::Azure => write!(f, "azure"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ServiceScalingPatchResponse.region`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceScalingPatchResponseRegion {
-    #[serde(rename = "ap-northeast-1")]
-    #[default]
-    Ap_northeast_1,
-    #[serde(rename = "ap-northeast-2")]
-    Ap_northeast_2,
-    #[serde(rename = "ap-south-1")]
-    Ap_south_1,
-    #[serde(rename = "ap-southeast-1")]
-    Ap_southeast_1,
-    #[serde(rename = "ap-southeast-2")]
-    Ap_southeast_2,
-    #[serde(rename = "ca-central-1")]
-    Ca_central_1,
-    #[serde(rename = "eu-central-1")]
-    Eu_central_1,
-    #[serde(rename = "eu-west-1")]
-    Eu_west_1,
-    #[serde(rename = "eu-west-2")]
-    Eu_west_2,
-    #[serde(rename = "il-central-1")]
-    Il_central_1,
-    #[serde(rename = "us-east-1")]
-    Us_east_1,
-    #[serde(rename = "us-east-2")]
-    Us_east_2,
-    #[serde(rename = "us-west-2")]
-    Us_west_2,
-    #[serde(rename = "us-east1")]
-    Us_east1,
-    #[serde(rename = "us-central1")]
-    Us_central1,
-    #[serde(rename = "europe-west2")]
-    Europe_west2,
-    #[serde(rename = "europe-west4")]
-    Europe_west4,
-    #[serde(rename = "asia-southeast1")]
-    Asia_southeast1,
-    #[serde(rename = "asia-northeast1")]
-    Asia_northeast1,
-    #[serde(rename = "eastus")]
-    Eastus,
-    #[serde(rename = "eastus2")]
-    Eastus2,
-    #[serde(rename = "westus3")]
-    Westus3,
-    #[serde(rename = "germanywestcentral")]
-    Germanywestcentral,
-    #[serde(rename = "centralus")]
-    Centralus,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceScalingPatchResponseRegion {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Ap_northeast_1 => write!(f, "ap-northeast-1"),
-            Self::Ap_northeast_2 => write!(f, "ap-northeast-2"),
-            Self::Ap_south_1 => write!(f, "ap-south-1"),
-            Self::Ap_southeast_1 => write!(f, "ap-southeast-1"),
-            Self::Ap_southeast_2 => write!(f, "ap-southeast-2"),
-            Self::Ca_central_1 => write!(f, "ca-central-1"),
-            Self::Eu_central_1 => write!(f, "eu-central-1"),
-            Self::Eu_west_1 => write!(f, "eu-west-1"),
-            Self::Eu_west_2 => write!(f, "eu-west-2"),
-            Self::Il_central_1 => write!(f, "il-central-1"),
-            Self::Us_east_1 => write!(f, "us-east-1"),
-            Self::Us_east_2 => write!(f, "us-east-2"),
-            Self::Us_west_2 => write!(f, "us-west-2"),
-            Self::Us_east1 => write!(f, "us-east1"),
-            Self::Us_central1 => write!(f, "us-central1"),
-            Self::Europe_west2 => write!(f, "europe-west2"),
-            Self::Europe_west4 => write!(f, "europe-west4"),
-            Self::Asia_southeast1 => write!(f, "asia-southeast1"),
-            Self::Asia_northeast1 => write!(f, "asia-northeast1"),
-            Self::Eastus => write!(f, "eastus"),
-            Self::Eastus2 => write!(f, "eastus2"),
-            Self::Westus3 => write!(f, "westus3"),
-            Self::Germanywestcentral => write!(f, "germanywestcentral"),
-            Self::Centralus => write!(f, "centralus"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ServiceScalingPatchResponse.releaseChannel`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceScalingPatchResponseReleasechannel {
-    #[serde(rename = "slow")]
-    #[default]
-    Slow,
-    #[serde(rename = "default")]
-    Default,
-    #[serde(rename = "fast")]
-    Fast,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceScalingPatchResponseReleasechannel {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Slow => write!(f, "slow"),
-            Self::Default => write!(f, "default"),
-            Self::Fast => write!(f, "fast"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ServiceScalingPatchResponse.state`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceScalingPatchResponseState {
-    #[serde(rename = "starting")]
-    #[default]
-    Starting,
-    #[serde(rename = "stopping")]
-    Stopping,
-    #[serde(rename = "terminating")]
-    Terminating,
-    #[serde(rename = "softdeleting")]
-    Softdeleting,
-    #[serde(rename = "awaking")]
-    Awaking,
-    #[serde(rename = "partially_running")]
-    Partially_running,
-    #[serde(rename = "provisioning")]
-    Provisioning,
-    #[serde(rename = "running")]
-    Running,
-    #[serde(rename = "stopped")]
-    Stopped,
-    #[serde(rename = "terminated")]
-    Terminated,
-    #[serde(rename = "softdeleted")]
-    Softdeleted,
-    #[serde(rename = "degraded")]
-    Degraded,
-    #[serde(rename = "failed")]
-    Failed,
-    #[serde(rename = "idle")]
-    Idle,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceScalingPatchResponseState {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Starting => write!(f, "starting"),
-            Self::Stopping => write!(f, "stopping"),
-            Self::Terminating => write!(f, "terminating"),
-            Self::Softdeleting => write!(f, "softdeleting"),
-            Self::Awaking => write!(f, "awaking"),
-            Self::Partially_running => write!(f, "partially_running"),
-            Self::Provisioning => write!(f, "provisioning"),
-            Self::Running => write!(f, "running"),
-            Self::Stopped => write!(f, "stopped"),
-            Self::Terminated => write!(f, "terminated"),
-            Self::Softdeleted => write!(f, "softdeleted"),
-            Self::Degraded => write!(f, "degraded"),
-            Self::Failed => write!(f, "failed"),
-            Self::Idle => write!(f, "idle"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ServiceScalingPatchResponse.tier`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceScalingPatchResponseTier {
-    #[serde(rename = "development")]
-    #[default]
-    Development,
-    #[serde(rename = "production")]
-    Production,
-    #[serde(rename = "dedicated_high_mem")]
-    Dedicated_high_mem,
-    #[serde(rename = "dedicated_high_cpu")]
-    Dedicated_high_cpu,
-    #[serde(rename = "dedicated_standard")]
-    Dedicated_standard,
-    #[serde(rename = "dedicated_standard_n2d_standard_4")]
-    Dedicated_standard_n2d_standard_4,
-    #[serde(rename = "dedicated_standard_n2d_standard_8")]
-    Dedicated_standard_n2d_standard_8,
-    #[serde(rename = "dedicated_standard_n2d_standard_32")]
-    Dedicated_standard_n2d_standard_32,
-    #[serde(rename = "dedicated_standard_n2d_standard_128")]
-    Dedicated_standard_n2d_standard_128,
-    #[serde(rename = "dedicated_standard_n2d_standard_32_16SSD")]
-    Dedicated_standard_n2d_standard_32_16SSD,
-    #[serde(rename = "dedicated_standard_n2d_standard_64_24SSD")]
-    Dedicated_standard_n2d_standard_64_24SSD,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceScalingPatchResponseTier {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Development => write!(f, "development"),
-            Self::Production => write!(f, "production"),
-            Self::Dedicated_high_mem => write!(f, "dedicated_high_mem"),
-            Self::Dedicated_high_cpu => write!(f, "dedicated_high_cpu"),
-            Self::Dedicated_standard => write!(f, "dedicated_standard"),
-            Self::Dedicated_standard_n2d_standard_4 => {
-                write!(f, "dedicated_standard_n2d_standard_4")
-            }
-            Self::Dedicated_standard_n2d_standard_8 => {
-                write!(f, "dedicated_standard_n2d_standard_8")
-            }
-            Self::Dedicated_standard_n2d_standard_32 => {
-                write!(f, "dedicated_standard_n2d_standard_32")
-            }
-            Self::Dedicated_standard_n2d_standard_128 => {
-                write!(f, "dedicated_standard_n2d_standard_128")
-            }
-            Self::Dedicated_standard_n2d_standard_32_16SSD => {
-                write!(f, "dedicated_standard_n2d_standard_32_16SSD")
-            }
-            Self::Dedicated_standard_n2d_standard_64_24SSD => {
-                write!(f, "dedicated_standard_n2d_standard_64_24SSD")
-            }
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `ServiceStatePatchRequest.command`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum ServiceStatePatchRequestCommand {
-    #[serde(rename = "start")]
-    #[default]
-    Start,
-    #[serde(rename = "stop")]
-    Stop,
-    #[serde(rename = "awake")]
-    Awake,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for ServiceStatePatchRequestCommand {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Start => write!(f, "start"),
-            Self::Stop => write!(f, "stop"),
-            Self::Awake => write!(f, "awake"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `UsageCostRecord.entityType`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum UsageCostRecordEntitytype {
-    #[serde(rename = "datawarehouse")]
-    #[default]
-    Datawarehouse,
-    #[serde(rename = "service")]
-    Service,
-    #[serde(rename = "clickpipe")]
-    Clickpipe,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for UsageCostRecordEntitytype {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Datawarehouse => write!(f, "datawarehouse"),
-            Self::Service => write!(f, "service"),
-            Self::Clickpipe => write!(f, "clickpipe"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `pgConfig.default_transaction_isolation`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum PgConfigDefaultTransactionIsolation {
-    #[serde(rename = "read committed")]
-    #[default]
-    Read_committed,
-    #[serde(rename = "repeatable read")]
-    Repeatable_read,
-    #[serde(rename = "serializable")]
-    Serializable,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for PgConfigDefaultTransactionIsolation {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Read_committed => write!(f, "read committed"),
-            Self::Repeatable_read => write!(f, "repeatable read"),
-            Self::Serializable => write!(f, "serializable"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `pgConfig.ssl_min_protocol_version`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum PgConfigSslMinProtocolVersion {
-    #[serde(rename = "TLSv1")]
-    #[default]
-    TlsV1,
-    #[serde(rename = "TLSv1.1")]
-    TlsV1_1,
-    #[serde(rename = "TLSv1.2")]
-    TlsV1_2,
-    #[serde(rename = "TLSv1.3")]
-    TlsV1_3,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for PgConfigSslMinProtocolVersion {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::TlsV1 => write!(f, "TLSv1"),
-            Self::TlsV1_1 => write!(f, "TLSv1.1"),
-            Self::TlsV1_2 => write!(f, "TLSv1.2"),
-            Self::TlsV1_3 => write!(f, "TLSv1.3"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// Inline enum for `pgConfig.wal_compression`.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub enum PgConfigWalCompression {
-    #[serde(rename = "off")]
-    #[default]
-    Off,
-    #[serde(rename = "on")]
-    On,
-    #[serde(rename = "lz4")]
-    Lz4,
-    #[serde(rename = "zstd")]
-    Zstd,
-    /// Catch-all for unknown or newly-added values.
-    #[serde(untagged)]
-    Unknown(String),
-}
-
-impl std::fmt::Display for PgConfigWalCompression {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Off => write!(f, "off"),
-            Self::On => write!(f, "on"),
-            Self::Lz4 => write!(f, "lz4"),
-            Self::Zstd => write!(f, "zstd"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// `BackupBucket` - one of multiple variants.
-///
-/// Dispatched on the `bucketProvider` field; see the `discriminated_union!`
-/// invocation below for the wire values.
-#[derive(Debug, Clone, PartialEq, Serialize)]
-#[serde(untagged)]
-pub enum BackupBucket {
-    AwsBackupBucket(AwsBackupBucket),
-    GcpBackupBucket(GcpBackupBucket),
-    AzureBackupBucket(AzureBackupBucket),
-    /// Catch-all for unknown or newly-added values.
-    ///
-    /// Holds the raw payload as `serde_json::Value` so it round-trips
-    /// losslessly; its `Display` emits the payload as compact JSON.
-    Unknown(serde_json::Value),
-}
-
-discriminated_union! {
-    BackupBucket, "bucketProvider" {
-        "AWS" => AwsBackupBucket,
-        "GCP" => GcpBackupBucket,
-        "AZURE" => AzureBackupBucket,
-    }
-}
-
-impl std::fmt::Display for BackupBucket {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::AwsBackupBucket(_) => write!(f, "AwsBackupBucket"),
-            Self::GcpBackupBucket(_) => write!(f, "GcpBackupBucket"),
-            Self::AzureBackupBucket(_) => write!(f, "AzureBackupBucket"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// `BackupBucketPatchRequest` - one of multiple variants.
-///
-/// Dispatched on the `bucketProvider` field; see the `discriminated_union!`
-/// invocation below for the wire values.
-#[derive(Debug, Clone, PartialEq, Serialize)]
-#[serde(untagged)]
-pub enum BackupBucketPatchRequest {
-    AwsBackupBucketPatchRequestV1(AwsBackupBucketPatchRequestV1),
-    GcpBackupBucketPatchRequestV1(GcpBackupBucketPatchRequestV1),
-    AzureBackupBucketPatchRequestV1(AzureBackupBucketPatchRequestV1),
-    /// Catch-all for unknown or newly-added values.
-    ///
-    /// Holds the raw payload as `serde_json::Value` so it round-trips
-    /// losslessly; its `Display` emits the payload as compact JSON.
-    Unknown(serde_json::Value),
-}
-
-discriminated_union! {
-    BackupBucketPatchRequest, "bucketProvider" {
-        "AWS" => AwsBackupBucketPatchRequestV1,
-        "GCP" => GcpBackupBucketPatchRequestV1,
-        "AZURE" => AzureBackupBucketPatchRequestV1,
-    }
-}
-
-impl std::fmt::Display for BackupBucketPatchRequest {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::AwsBackupBucketPatchRequestV1(_) => write!(f, "AwsBackupBucketPatchRequestV1"),
-            Self::GcpBackupBucketPatchRequestV1(_) => write!(f, "GcpBackupBucketPatchRequestV1"),
-            Self::AzureBackupBucketPatchRequestV1(_) => {
-                write!(f, "AzureBackupBucketPatchRequestV1")
-            }
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// `BackupBucketPostRequest` - one of multiple variants.
-///
-/// Dispatched on the `bucketProvider` field; see the `discriminated_union!`
-/// invocation below for the wire values.
-#[derive(Debug, Clone, PartialEq, Serialize)]
-#[serde(untagged)]
-pub enum BackupBucketPostRequest {
-    AwsBackupBucketPostRequestV1(AwsBackupBucketPostRequestV1),
-    GcpBackupBucketPostRequestV1(GcpBackupBucketPostRequestV1),
-    AzureBackupBucketPostRequestV1(AzureBackupBucketPostRequestV1),
-    /// Catch-all for unknown or newly-added values.
-    ///
-    /// Holds the raw payload as `serde_json::Value` so it round-trips
-    /// losslessly; its `Display` emits the payload as compact JSON.
-    Unknown(serde_json::Value),
-}
-
-discriminated_union! {
-    BackupBucketPostRequest, "bucketProvider" {
-        "AWS" => AwsBackupBucketPostRequestV1,
-        "GCP" => GcpBackupBucketPostRequestV1,
-        "AZURE" => AzureBackupBucketPostRequestV1,
-    }
-}
-
-impl std::fmt::Display for BackupBucketPostRequest {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::AwsBackupBucketPostRequestV1(_) => write!(f, "AwsBackupBucketPostRequestV1"),
-            Self::GcpBackupBucketPostRequestV1(_) => write!(f, "GcpBackupBucketPostRequestV1"),
-            Self::AzureBackupBucketPostRequestV1(_) => write!(f, "AzureBackupBucketPostRequestV1"),
-            Self::Unknown(s) => write!(f, "{s}"),
-        }
-    }
-}
-
-/// `BackupBucketProperties` - one of multiple variants.
-///
-/// Dispatched on the `bucketProvider` field; see the `discriminated_union!`
-/// invocation below for the wire values.
-#[derive(Debug, Clone, PartialEq, Serialize)]
-#[serde(untagged)]
-pub enum BackupBucketProperties {
-    AwsBackupBucketProperties(AwsBackupBucketProperties),
-    GcpBackupBucketProperties(GcpBackupBucketProperties),
-    AzureBackupBucketProperties(AzureBackupBucketProperties),
-    /// Catch-all for unknown or newly-added values.
-    ///
-    /// Holds the raw payload as `serde_json::Value` so it round-trips
-    /// losslessly; its `Display` emits the payload as compact JSON.
-    Unknown(serde_json::Value),
-}
-
-discriminated_union! {
-    BackupBucketProperties, "bucketProvider" {
-        "AWS" => AwsBackupBucketProperties,
-        "GCP" => GcpBackupBucketProperties,
-        "AZURE" => AzureBackupBucketProperties,
-    }
-}
-
-impl std::fmt::Display for BackupBucketProperties {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::AwsBackupBucketProperties(_) => write!(f, "AwsBackupBucketProperties"),
-            Self::GcpBackupBucketProperties(_) => write!(f, "GcpBackupBucketProperties"),
-            Self::AzureBackupBucketProperties(_) => write!(f, "AzureBackupBucketProperties"),
             Self::Unknown(s) => write!(f, "{s}"),
         }
     }
@@ -8135,211 +6009,11 @@ pub type ClickStackWebhookInputHeaders = std::collections::BTreeMap<String, Stri
 /// Type alias for `ClickStackWebhookInputQueryParams`.
 pub type ClickStackWebhookInputQueryParams = std::collections::BTreeMap<String, String>;
 
-/// Type alias for `pgCreatedAtProperty`.
-pub type PgCreatedAtProperty = chrono::DateTime<chrono::Utc>;
-
-/// Type alias for `pgIdProperty`.
-pub type PgIdProperty = uuid::Uuid;
-
-/// Type alias for `pgIsPrimaryProperty`.
-pub type PgIsPrimaryProperty = bool;
-
-/// Type alias for `pgNameProperty`.
-pub type PgNameProperty = String;
-
-/// Type alias for `pgPassword`.
-pub type PgPassword = String;
-
-/// Type alias for `pgPitrRestoreTargetProperty`.
-pub type PgPitrRestoreTargetProperty = chrono::DateTime<chrono::Utc>;
-
-/// Type alias for `pgRegion`.
-pub type PgRegion = String;
-
-/// Type alias for `pgStorageSize`.
-pub type PgStorageSize = i64;
-
-/// Type alias for `pgTags`.
-pub type PgTags = Vec<ResourceTagsV1>;
-
-/// Type alias for `pgTags` in response position, over
-/// [`ResourceTagsV1Response`].
-pub type PgTagsResponse = Vec<ResourceTagsV1Response>;
-
-/// `AwsBackupBucket` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct AwsBackupBucket {
-    #[serde(rename = "bucketPath", skip_serializing_if = "Option::is_none")]
-    pub bucket_path: Option<String>,
-    #[serde(rename = "bucketProvider", skip_serializing_if = "Option::is_none")]
-    pub bucket_provider: Option<AwsBackupBucketBucketprovider>,
-    #[serde(rename = "iamRoleArn", skip_serializing_if = "Option::is_none")]
-    pub iam_role_arn: Option<String>,
-    #[serde(rename = "iamRoleSessionName", skip_serializing_if = "Option::is_none")]
-    pub iam_role_session_name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<uuid::Uuid>,
-}
-
-/// `AwsBackupBucketPatchRequestV1` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct AwsBackupBucketPatchRequestV1 {
-    #[serde(rename = "bucketPath")]
-    pub bucket_path: String,
-    #[serde(rename = "bucketProvider")]
-    pub bucket_provider: AwsBackupBucketPatchRequestV1Bucketprovider,
-    #[serde(rename = "iamRoleArn")]
-    pub iam_role_arn: String,
-    #[serde(rename = "iamRoleSessionName", skip_serializing_if = "Option::is_none")]
-    pub iam_role_session_name: Option<String>,
-}
-
-/// `AwsBackupBucketPostRequestV1` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct AwsBackupBucketPostRequestV1 {
-    #[serde(rename = "bucketPath")]
-    pub bucket_path: String,
-    #[serde(rename = "bucketProvider")]
-    pub bucket_provider: AwsBackupBucketPostRequestV1Bucketprovider,
-    #[serde(rename = "iamRoleArn")]
-    pub iam_role_arn: String,
-    #[serde(rename = "iamRoleSessionName")]
-    pub iam_role_session_name: String,
-}
-
-/// `AwsBackupBucketProperties` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct AwsBackupBucketProperties {
-    #[serde(rename = "bucketPath")]
-    pub bucket_path: String,
-    #[serde(rename = "bucketProvider")]
-    pub bucket_provider: AwsBackupBucketPropertiesBucketprovider,
-    #[serde(rename = "iamRoleArn")]
-    pub iam_role_arn: String,
-    #[serde(rename = "iamRoleSessionName")]
-    pub iam_role_session_name: String,
-}
-
-/// `AzureBackupBucket` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct AzureBackupBucket {
-    #[serde(rename = "bucketProvider", skip_serializing_if = "Option::is_none")]
-    pub bucket_provider: Option<AzureBackupBucketBucketprovider>,
-    #[serde(rename = "containerName", skip_serializing_if = "Option::is_none")]
-    pub container_name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<uuid::Uuid>,
-}
-
-/// `AzureBackupBucketPatchRequestV1` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct AzureBackupBucketPatchRequestV1 {
-    #[serde(rename = "bucketProvider")]
-    pub bucket_provider: AzureBackupBucketPatchRequestV1Bucketprovider,
-    #[serde(rename = "connectionString")]
-    pub connection_string: String,
-    #[serde(rename = "containerName")]
-    pub container_name: String,
-}
-
-/// `AzureBackupBucketPostRequestV1` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct AzureBackupBucketPostRequestV1 {
-    #[serde(rename = "bucketProvider")]
-    pub bucket_provider: AzureBackupBucketPostRequestV1Bucketprovider,
-    #[serde(rename = "connectionString")]
-    pub connection_string: String,
-    #[serde(rename = "containerName")]
-    pub container_name: String,
-}
-
-/// `AzureBackupBucketProperties` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct AzureBackupBucketProperties {
-    #[serde(rename = "bucketProvider")]
-    pub bucket_provider: AzureBackupBucketPropertiesBucketprovider,
-    #[serde(rename = "containerName")]
-    pub container_name: String,
-}
-
 /// `AzureEventHub` from the ClickHouse Cloud API.
 #[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
 pub struct AzureEventHub {
     #[serde(rename = "connectionString")]
     pub connection_string: String,
-}
-
-/// `Backup` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct Backup {
-    #[serde(rename = "backupName", skip_serializing_if = "Option::is_none")]
-    pub backup_name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub bucket: Option<serde_json::Value>,
-    #[serde(rename = "durationInSeconds", skip_serializing_if = "Option::is_none")]
-    pub duration_in_seconds: Option<f64>,
-    #[serde(rename = "finishedAt", skip_serializing_if = "Option::is_none")]
-    pub finished_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<uuid::Uuid>,
-    #[serde(rename = "serviceId", skip_serializing_if = "Option::is_none")]
-    pub service_id: Option<String>,
-    #[serde(rename = "sizeInBytes", skip_serializing_if = "Option::is_none")]
-    pub size_in_bytes: Option<f64>,
-    #[serde(rename = "startedAt", skip_serializing_if = "Option::is_none")]
-    pub started_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub status: Option<BackupStatus>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub r#type: Option<BackupType>,
-}
-
-/// `BackupConfiguration` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct BackupConfiguration {
-    #[serde(
-        rename = "backupPeriodInHours",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub backup_period_in_hours: Option<f64>,
-    #[serde(
-        rename = "backupRetentionPeriodInHours",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub backup_retention_period_in_hours: Option<f64>,
-    #[serde(rename = "backupStartTime", skip_serializing_if = "Option::is_none")]
-    pub backup_start_time: Option<String>,
-}
-
-/// `BackupConfigurationPatchRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct BackupConfigurationPatchRequest {
-    #[serde(
-        rename = "backupPeriodInHours",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub backup_period_in_hours: Option<f64>,
-    #[serde(
-        rename = "backupRetentionPeriodInHours",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub backup_retention_period_in_hours: Option<f64>,
-    #[serde(rename = "backupStartTime", skip_serializing_if = "Option::is_none")]
-    pub backup_start_time: Option<String>,
-}
-
-/// `BasePostgresService` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct BasePostgresService {
-    #[serde(rename = "haType")]
-    pub ha_type: PgHaType,
-    pub name: PgNameProperty,
-    #[serde(rename = "postgresVersion")]
-    pub postgres_version: PgVersion,
-    pub provider: PgProvider,
-    pub region: PgRegion,
-    pub size: PgSize,
-    pub tags: PgTags,
 }
 
 /// `ClickPipe` from the ClickHouse Cloud API.
@@ -12954,48 +10628,6 @@ pub struct CreateReversePrivateEndpoint {
     pub vpc_resource_share_arn: Option<String>,
 }
 
-/// `CurrentScaling` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct CurrentScaling {
-    #[serde(rename = "activeEntryId", skip_serializing_if = "Option::is_none")]
-    pub active_entry_id: Option<uuid::Uuid>,
-    #[serde(
-        rename = "effectiveAutoscalingMode",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub effective_autoscaling_mode: Option<CurrentScalingEffectiveautoscalingmode>,
-    #[serde(
-        rename = "effectiveIdleScaling",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub effective_idle_scaling: Option<bool>,
-    #[serde(
-        rename = "effectiveIdleTimeoutMinutes",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub effective_idle_timeout_minutes: Option<i64>,
-    #[serde(
-        rename = "effectiveMaxReplicaMemoryGb",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub effective_max_replica_memory_gb: Option<f64>,
-    #[serde(
-        rename = "effectiveMaxReplicas",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub effective_max_replicas: Option<i64>,
-    #[serde(
-        rename = "effectiveMinReplicaMemoryGb",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub effective_min_replica_memory_gb: Option<f64>,
-    #[serde(
-        rename = "effectiveMinReplicas",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub effective_min_replicas: Option<i64>,
-}
-
 /// `CustomPrivateDnsMapping` from the ClickHouse Cloud API.
 #[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
 pub struct CustomPrivateDnsMapping {
@@ -13013,93 +10645,6 @@ pub struct CustomPrivateDnsMapping {
 pub struct CustomPrivateDnsMappingResponse {
     #[serde(rename = "privateDnsName", skip_serializing_if = "Option::is_none")]
     pub private_dns_name: Option<String>,
-}
-
-/// `GcpBackupBucket` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct GcpBackupBucket {
-    #[serde(rename = "accessKeyId", skip_serializing_if = "Option::is_none")]
-    pub access_key_id: Option<String>,
-    #[serde(rename = "bucketPath", skip_serializing_if = "Option::is_none")]
-    pub bucket_path: Option<String>,
-    #[serde(rename = "bucketProvider", skip_serializing_if = "Option::is_none")]
-    pub bucket_provider: Option<GcpBackupBucketBucketprovider>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<uuid::Uuid>,
-}
-
-/// `GcpBackupBucketPatchRequestV1` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct GcpBackupBucketPatchRequestV1 {
-    #[serde(rename = "accessKeyId")]
-    pub access_key_id: String,
-    #[serde(rename = "bucketPath")]
-    pub bucket_path: String,
-    #[serde(rename = "bucketProvider")]
-    pub bucket_provider: GcpBackupBucketPatchRequestV1Bucketprovider,
-    #[serde(rename = "secretAccessKey")]
-    pub secret_access_key: String,
-}
-
-/// `GcpBackupBucketPostRequestV1` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct GcpBackupBucketPostRequestV1 {
-    #[serde(rename = "accessKeyId")]
-    pub access_key_id: String,
-    #[serde(rename = "bucketPath")]
-    pub bucket_path: String,
-    #[serde(rename = "bucketProvider")]
-    pub bucket_provider: GcpBackupBucketPostRequestV1Bucketprovider,
-    #[serde(rename = "secretAccessKey")]
-    pub secret_access_key: String,
-}
-
-/// `GcpBackupBucketProperties` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct GcpBackupBucketProperties {
-    #[serde(rename = "accessKeyId")]
-    pub access_key_id: String,
-    #[serde(rename = "bucketPath")]
-    pub bucket_path: String,
-    #[serde(rename = "bucketProvider")]
-    pub bucket_provider: GcpBackupBucketPropertiesBucketprovider,
-}
-
-/// `InstancePrivateEndpoint` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct InstancePrivateEndpoint {
-    #[serde(rename = "cloudProvider", skip_serializing_if = "Option::is_none")]
-    pub cloud_provider: Option<InstancePrivateEndpointCloudprovider>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub region: Option<InstancePrivateEndpointRegion>,
-}
-
-/// `InstancePrivateEndpointsPatch` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct InstancePrivateEndpointsPatch {
-    pub add: Vec<String>,
-    pub remove: Vec<String>,
-}
-
-/// `InstanceServiceQueryApiEndpointsPostRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct InstanceServiceQueryApiEndpointsPostRequest {
-    #[serde(rename = "allowedOrigins")]
-    pub allowed_origins: String,
-    #[serde(rename = "openApiKeys")]
-    pub open_api_keys: Vec<String>,
-    pub roles: Vec<String>,
-}
-
-/// `InstanceTagsPatch` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct InstanceTagsPatch {
-    pub add: Vec<ResourceTagsV1>,
-    pub remove: Vec<ResourceTagsV1>,
 }
 
 /// `MskIamUser` from the ClickHouse Cloud API.
@@ -13124,360 +10669,6 @@ pub struct MutualTLS {
 pub struct PLAIN {
     pub password: String,
     pub username: String,
-}
-
-/// `PostgresService` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresService {
-    #[serde(rename = "connectionString", skip_serializing_if = "Option::is_none")]
-    pub connection_string: Option<String>,
-    #[serde(rename = "createdAt", skip_serializing_if = "Option::is_none")]
-    pub created_at: Option<PgCreatedAtProperty>,
-    #[serde(rename = "haType", skip_serializing_if = "Option::is_none")]
-    pub ha_type: Option<PgHaType>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub hostname: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<PgIdProperty>,
-    #[serde(rename = "isPrimary", skip_serializing_if = "Option::is_none")]
-    pub is_primary: Option<PgIsPrimaryProperty>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<PgNameProperty>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub password: Option<String>,
-    #[serde(rename = "postgresVersion", skip_serializing_if = "Option::is_none")]
-    pub postgres_version: Option<PgVersion>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub provider: Option<PgProvider>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub region: Option<PgRegion>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub size: Option<PgSize>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub state: Option<PgStateProperty>,
-    #[serde(rename = "storageSize", skip_serializing_if = "Option::is_none")]
-    pub storage_size: Option<PgStorageSize>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tags: Option<PgTagsResponse>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub username: Option<String>,
-}
-
-/// `PostgresServiceListItem` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresServiceListItem {
-    #[serde(rename = "createdAt", skip_serializing_if = "Option::is_none")]
-    pub created_at: Option<PgCreatedAtProperty>,
-    #[serde(rename = "haType", skip_serializing_if = "Option::is_none")]
-    pub ha_type: Option<PgHaType>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<PgIdProperty>,
-    #[serde(rename = "isPrimary", skip_serializing_if = "Option::is_none")]
-    pub is_primary: Option<PgIsPrimaryProperty>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<PgNameProperty>,
-    #[serde(rename = "postgresVersion", skip_serializing_if = "Option::is_none")]
-    pub postgres_version: Option<PgVersion>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub provider: Option<PgProvider>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub region: Option<PgRegion>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub size: Option<PgSize>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub state: Option<PgStateProperty>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tags: Option<PgTagsResponse>,
-}
-
-/// `PostgresServicePasswordResource` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresServicePasswordResource {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub password: Option<String>,
-}
-
-/// `PostgresServicePatchRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresServicePatchRequest {
-    #[serde(rename = "haType", skip_serializing_if = "Option::is_none")]
-    pub ha_type: Option<PgHaType>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<PgNameProperty>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub size: Option<PgSize>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tags: Option<PgTags>,
-}
-
-/// `PostgresServicePostRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresServicePostRequest {
-    #[serde(rename = "haType", skip_serializing_if = "Option::is_none")]
-    pub ha_type: Option<PgHaType>,
-    pub name: PgNameProperty,
-    #[serde(rename = "pgBouncerConfig", skip_serializing_if = "Option::is_none")]
-    pub pg_bouncer_config: Option<PgBouncerConfig>,
-    #[serde(rename = "pgConfig", skip_serializing_if = "Option::is_none")]
-    pub pg_config: Option<PgConfig>,
-    #[serde(rename = "postgresVersion", skip_serializing_if = "Option::is_none")]
-    pub postgres_version: Option<PgVersion>,
-    pub provider: PgProvider,
-    pub region: PgRegion,
-    pub size: PgSize,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tags: Option<PgTags>,
-}
-
-/// `PostgresServiceReadReplicaRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresServiceReadReplicaRequest {
-    pub name: PgNameProperty,
-    #[serde(rename = "pgBouncerConfig", skip_serializing_if = "Option::is_none")]
-    pub pg_bouncer_config: Option<PgBouncerConfig>,
-    #[serde(rename = "pgConfig", skip_serializing_if = "Option::is_none")]
-    pub pg_config: Option<PgConfig>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tags: Option<PgTags>,
-}
-
-/// `PostgresServiceRestoreRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresServiceRestoreRequest {
-    pub name: PgNameProperty,
-    #[serde(rename = "pgBouncerConfig", skip_serializing_if = "Option::is_none")]
-    pub pg_bouncer_config: Option<PgBouncerConfig>,
-    #[serde(rename = "pgConfig", skip_serializing_if = "Option::is_none")]
-    pub pg_config: Option<PgConfig>,
-    #[serde(rename = "restoreTarget")]
-    pub restore_target: PgPitrRestoreTargetProperty,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tags: Option<PgTags>,
-}
-
-/// `PostgresServiceSetPassword` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresServiceSetPassword {
-    pub password: PgPassword,
-}
-
-/// `PostgresServiceSetState` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresServiceSetState {
-    pub command: PostgresServiceSetStateCommand,
-}
-
-/// `PostgresMetricDataPoint` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresMetricDataPoint {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub timestamp: Option<i64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub value: Option<f64>,
-}
-
-/// `PostgresMetricSeries` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresMetricSeries {
-    #[serde(rename = "dataPoints", skip_serializing_if = "Option::is_none")]
-    pub data_points: Option<Vec<PostgresMetricDataPoint>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub label: Option<String>,
-}
-
-/// `PostgresMetric` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresMetric {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub key: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub series: Option<Vec<PostgresMetricSeries>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub unit: Option<String>,
-}
-
-/// `PostgresMetrics` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresMetrics {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub metrics: Option<Vec<PostgresMetric>>,
-}
-
-/// `PostgresQueryExecution` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresQueryExecution {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub app: Option<String>,
-    #[serde(rename = "cpuSysTimeUs", skip_serializing_if = "Option::is_none")]
-    pub cpu_sys_time_us: Option<i64>,
-    #[serde(rename = "cpuUserTimeUs", skip_serializing_if = "Option::is_none")]
-    pub cpu_user_time_us: Option<i64>,
-    #[serde(rename = "dbName", skip_serializing_if = "Option::is_none")]
-    pub db_name: Option<String>,
-    #[serde(rename = "dbOperation", skip_serializing_if = "Option::is_none")]
-    pub db_operation: Option<String>,
-    #[serde(rename = "dbUser", skip_serializing_if = "Option::is_none")]
-    pub db_user: Option<String>,
-    #[serde(rename = "durationUs", skip_serializing_if = "Option::is_none")]
-    pub duration_us: Option<i64>,
-    #[serde(rename = "errElevel", skip_serializing_if = "Option::is_none")]
-    pub err_elevel: Option<i64>,
-    #[serde(rename = "errMessage", skip_serializing_if = "Option::is_none")]
-    pub err_message: Option<String>,
-    #[serde(rename = "errSqlstate", skip_serializing_if = "Option::is_none")]
-    pub err_sqlstate: Option<String>,
-    #[serde(rename = "jitDeformTimeUs", skip_serializing_if = "Option::is_none")]
-    pub jit_deform_time_us: Option<i64>,
-    #[serde(rename = "jitEmissionTimeUs", skip_serializing_if = "Option::is_none")]
-    pub jit_emission_time_us: Option<i64>,
-    #[serde(rename = "jitFunctions", skip_serializing_if = "Option::is_none")]
-    pub jit_functions: Option<i64>,
-    #[serde(
-        rename = "jitGenerationTimeUs",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub jit_generation_time_us: Option<i64>,
-    #[serde(rename = "jitInliningTimeUs", skip_serializing_if = "Option::is_none")]
-    pub jit_inlining_time_us: Option<i64>,
-    #[serde(
-        rename = "jitOptimizationTimeUs",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub jit_optimization_time_us: Option<i64>,
-    #[serde(rename = "localBlksDirtied", skip_serializing_if = "Option::is_none")]
-    pub local_blks_dirtied: Option<i64>,
-    #[serde(rename = "localBlksHit", skip_serializing_if = "Option::is_none")]
-    pub local_blks_hit: Option<i64>,
-    #[serde(rename = "localBlksRead", skip_serializing_if = "Option::is_none")]
-    pub local_blks_read: Option<i64>,
-    #[serde(rename = "localBlksWritten", skip_serializing_if = "Option::is_none")]
-    pub local_blks_written: Option<i64>,
-    #[serde(
-        rename = "parallelWorkersLaunched",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub parallel_workers_launched: Option<i64>,
-    #[serde(
-        rename = "parallelWorkersPlanned",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub parallel_workers_planned: Option<i64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub pid: Option<String>,
-    #[serde(rename = "queryId", skip_serializing_if = "Option::is_none")]
-    pub query_id: Option<String>,
-    #[serde(rename = "queryText", skip_serializing_if = "Option::is_none")]
-    pub query_text: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub rows: Option<i64>,
-    #[serde(rename = "serverRole", skip_serializing_if = "Option::is_none")]
-    pub server_role: Option<String>,
-    #[serde(
-        rename = "sharedBlkReadTimeUs",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub shared_blk_read_time_us: Option<i64>,
-    #[serde(
-        rename = "sharedBlkWriteTimeUs",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub shared_blk_write_time_us: Option<i64>,
-    #[serde(rename = "sharedBlksDirtied", skip_serializing_if = "Option::is_none")]
-    pub shared_blks_dirtied: Option<i64>,
-    #[serde(rename = "sharedBlksHit", skip_serializing_if = "Option::is_none")]
-    pub shared_blks_hit: Option<i64>,
-    #[serde(rename = "sharedBlksRead", skip_serializing_if = "Option::is_none")]
-    pub shared_blks_read: Option<i64>,
-    #[serde(rename = "sharedBlksWritten", skip_serializing_if = "Option::is_none")]
-    pub shared_blks_written: Option<i64>,
-    #[serde(rename = "spanId", skip_serializing_if = "Option::is_none")]
-    pub span_id: Option<String>,
-    #[serde(rename = "tempBlkReadTimeUs", skip_serializing_if = "Option::is_none")]
-    pub temp_blk_read_time_us: Option<i64>,
-    #[serde(rename = "tempBlkWriteTimeUs", skip_serializing_if = "Option::is_none")]
-    pub temp_blk_write_time_us: Option<i64>,
-    #[serde(rename = "tempBlksRead", skip_serializing_if = "Option::is_none")]
-    pub temp_blks_read: Option<i64>,
-    #[serde(rename = "tempBlksWritten", skip_serializing_if = "Option::is_none")]
-    pub temp_blks_written: Option<i64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub timestamp: Option<String>,
-    #[serde(rename = "traceId", skip_serializing_if = "Option::is_none")]
-    pub trace_id: Option<String>,
-    #[serde(rename = "walBytes", skip_serializing_if = "Option::is_none")]
-    pub wal_bytes: Option<i64>,
-    #[serde(rename = "walFpi", skip_serializing_if = "Option::is_none")]
-    pub wal_fpi: Option<i64>,
-    #[serde(rename = "walRecords", skip_serializing_if = "Option::is_none")]
-    pub wal_records: Option<i64>,
-}
-
-/// `PostgresSlowQueryPattern` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresSlowQueryPattern {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub app: Option<String>,
-    #[serde(rename = "avgDurationUs", skip_serializing_if = "Option::is_none")]
-    pub avg_duration_us: Option<i64>,
-    #[serde(rename = "callCount", skip_serializing_if = "Option::is_none")]
-    pub call_count: Option<i64>,
-    #[serde(rename = "dbName", skip_serializing_if = "Option::is_none")]
-    pub db_name: Option<String>,
-    #[serde(rename = "dbOperation", skip_serializing_if = "Option::is_none")]
-    pub db_operation: Option<String>,
-    #[serde(rename = "dbUser", skip_serializing_if = "Option::is_none")]
-    pub db_user: Option<String>,
-    #[serde(rename = "errorCount", skip_serializing_if = "Option::is_none")]
-    pub error_count: Option<i64>,
-    #[serde(rename = "maxDurationUs", skip_serializing_if = "Option::is_none")]
-    pub max_duration_us: Option<i64>,
-    #[serde(rename = "p50DurationUs", skip_serializing_if = "Option::is_none")]
-    pub p50_duration_us: Option<i64>,
-    #[serde(rename = "p95DurationUs", skip_serializing_if = "Option::is_none")]
-    pub p95_duration_us: Option<i64>,
-    #[serde(rename = "p99DurationUs", skip_serializing_if = "Option::is_none")]
-    pub p99_duration_us: Option<i64>,
-    #[serde(rename = "queryId", skip_serializing_if = "Option::is_none")]
-    pub query_id: Option<String>,
-    #[serde(rename = "queryText", skip_serializing_if = "Option::is_none")]
-    pub query_text: Option<String>,
-    #[serde(rename = "totalCpuTimeUs", skip_serializing_if = "Option::is_none")]
-    pub total_cpu_time_us: Option<i64>,
-    #[serde(rename = "totalDurationUs", skip_serializing_if = "Option::is_none")]
-    pub total_duration_us: Option<i64>,
-    #[serde(rename = "totalRows", skip_serializing_if = "Option::is_none")]
-    pub total_rows: Option<i64>,
-    #[serde(rename = "totalSharedBlksHit", skip_serializing_if = "Option::is_none")]
-    pub total_shared_blks_hit: Option<i64>,
-    #[serde(
-        rename = "totalSharedBlksRead",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub total_shared_blks_read: Option<i64>,
-    #[serde(rename = "totalWalBytes", skip_serializing_if = "Option::is_none")]
-    pub total_wal_bytes: Option<i64>,
-}
-
-/// `PostgresSlowQueryPatternDetail` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresSlowQueryPatternDetail {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub aggregate: Option<PostgresSlowQueryPattern>,
-    #[serde(rename = "recentExecutions", skip_serializing_if = "Option::is_none")]
-    pub recent_executions: Option<Vec<PostgresQueryExecution>>,
-}
-
-/// `PrivateEndpointConfig` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PrivateEndpointConfig {
-    #[serde(rename = "endpointServiceId", skip_serializing_if = "Option::is_none")]
-    pub endpoint_service_id: Option<String>,
-    #[serde(rename = "privateDnsHostname", skip_serializing_if = "Option::is_none")]
-    pub private_dns_hostname: Option<String>,
 }
 
 /// `ReversePrivateEndpoint` from the ClickHouse Cloud API.
@@ -13530,604 +10721,11 @@ pub struct ReversePrivateEndpoint {
     pub vpc_resource_share_arn: Option<String>,
 }
 
-/// `ScalingSchedule` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScalingSchedule {
-    #[serde(rename = "activeEntryId", skip_serializing_if = "Option::is_none")]
-    pub active_entry_id: Option<uuid::Uuid>,
-    #[serde(rename = "baseConfig", skip_serializing_if = "Option::is_none")]
-    pub base_config: Option<ScalingScheduleBaseConfig>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub entries: Option<Vec<ScalingScheduleEntry>>,
-}
-
-/// `ScalingScheduleBaseConfig` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScalingScheduleBaseConfig {
-    #[serde(rename = "autoscalingMode", skip_serializing_if = "Option::is_none")]
-    pub autoscaling_mode: Option<AutoscalingMode>,
-    #[serde(rename = "idleScaling", skip_serializing_if = "Option::is_none")]
-    pub idle_scaling: Option<bool>,
-    #[serde(rename = "idleTimeoutMinutes", skip_serializing_if = "Option::is_none")]
-    pub idle_timeout_minutes: Option<i64>,
-    #[serde(rename = "maxReplicaMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub max_replica_memory_gb: Option<f64>,
-    #[serde(rename = "maxReplicas", skip_serializing_if = "Option::is_none")]
-    pub max_replicas: Option<i64>,
-    #[serde(rename = "minReplicaMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub min_replica_memory_gb: Option<f64>,
-    #[serde(rename = "minReplicas", skip_serializing_if = "Option::is_none")]
-    pub min_replicas: Option<i64>,
-}
-
-/// `ScalingScheduleEntry` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScalingScheduleEntry {
-    #[serde(rename = "autoscalingMode", skip_serializing_if = "Option::is_none")]
-    pub autoscaling_mode: Option<AutoscalingMode>,
-    #[serde(rename = "endHourUtc", skip_serializing_if = "Option::is_none")]
-    pub end_hour_utc: Option<i64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<uuid::Uuid>,
-    #[serde(rename = "idleScaling", skip_serializing_if = "Option::is_none")]
-    pub idle_scaling: Option<bool>,
-    #[serde(rename = "idleTimeoutMinutes", skip_serializing_if = "Option::is_none")]
-    pub idle_timeout_minutes: Option<i64>,
-    #[serde(rename = "isActiveNow", skip_serializing_if = "Option::is_none")]
-    pub is_active_now: Option<bool>,
-    #[serde(rename = "maxReplicaMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub max_replica_memory_gb: Option<f64>,
-    #[serde(rename = "maxReplicas", skip_serializing_if = "Option::is_none")]
-    pub max_replicas: Option<i64>,
-    #[serde(rename = "minReplicaMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub min_replica_memory_gb: Option<f64>,
-    #[serde(rename = "minReplicas", skip_serializing_if = "Option::is_none")]
-    pub min_replicas: Option<i64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(rename = "startHourUtc", skip_serializing_if = "Option::is_none")]
-    pub start_hour_utc: Option<i64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub weekdays: Option<Vec<i64>>,
-}
-
-/// `ScalingScheduleEntryRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScalingScheduleEntryRequest {
-    #[serde(rename = "autoscalingMode", skip_serializing_if = "Option::is_none")]
-    pub autoscaling_mode: Option<AutoscalingMode>,
-    #[serde(rename = "endHourUtc")]
-    pub end_hour_utc: i64,
-    #[serde(rename = "idleScaling", skip_serializing_if = "Option::is_none")]
-    pub idle_scaling: Option<bool>,
-    #[serde(rename = "idleTimeoutMinutes", skip_serializing_if = "Option::is_none")]
-    pub idle_timeout_minutes: Option<i64>,
-    #[serde(rename = "maxReplicaMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub max_replica_memory_gb: Option<f64>,
-    #[serde(rename = "maxReplicas", skip_serializing_if = "Option::is_none")]
-    pub max_replicas: Option<i64>,
-    #[serde(rename = "minReplicaMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub min_replica_memory_gb: Option<f64>,
-    #[serde(rename = "minReplicas", skip_serializing_if = "Option::is_none")]
-    pub min_replicas: Option<i64>,
-    pub name: String,
-    #[serde(rename = "numReplicas", skip_serializing_if = "Option::is_none")]
-    pub num_replicas: Option<i64>,
-    #[serde(rename = "startHourUtc")]
-    pub start_hour_utc: i64,
-    pub weekdays: Vec<i64>,
-}
-
-/// `ScalingSchedulePostRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ScalingSchedulePostRequest {
-    pub entries: Vec<ScalingScheduleEntryRequest>,
-}
-
-/// `ServicPrivateEndpointePostRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServicPrivateEndpointePostRequest {
-    pub description: String,
-    pub id: String,
-}
-
-/// `Service` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct Service {
-    #[serde(
-        rename = "availablePrivateEndpointIds",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub available_private_endpoint_ids: Option<Vec<String>>,
-    #[serde(rename = "autoscalingMode", skip_serializing_if = "Option::is_none")]
-    pub autoscaling_mode: Option<AutoscalingMode>,
-    #[serde(rename = "byocId", skip_serializing_if = "Option::is_none")]
-    pub byoc_id: Option<String>,
-    #[serde(rename = "clickhouseVersion", skip_serializing_if = "Option::is_none")]
-    pub clickhouse_version: Option<String>,
-    #[serde(rename = "complianceType", skip_serializing_if = "Option::is_none")]
-    pub compliance_type: Option<ServiceCompliancetype>,
-    #[serde(rename = "createdAt", skip_serializing_if = "Option::is_none")]
-    pub created_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(rename = "currentScaling", skip_serializing_if = "Option::is_none")]
-    pub current_scaling: Option<CurrentScaling>,
-    #[serde(rename = "dataWarehouseId", skip_serializing_if = "Option::is_none")]
-    pub data_warehouse_id: Option<String>,
-    #[serde(rename = "enableCoreDumps", skip_serializing_if = "Option::is_none")]
-    pub enable_core_dumps: Option<bool>,
-    #[serde(
-        rename = "encryptionAssumedRoleIdentifier",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub encryption_assumed_role_identifier: Option<String>,
-    #[serde(rename = "encryptionKey", skip_serializing_if = "Option::is_none")]
-    pub encryption_key: Option<String>,
-    #[serde(rename = "encryptionRoleId", skip_serializing_if = "Option::is_none")]
-    pub encryption_role_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub endpoints: Option<Vec<ServiceEndpoint>>,
-    #[serde(
-        rename = "hasTransparentDataEncryption",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub has_transparent_data_encryption: Option<bool>,
-    #[serde(rename = "iamRole", skip_serializing_if = "Option::is_none")]
-    pub iam_role: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<uuid::Uuid>,
-    #[serde(rename = "idleScaling", skip_serializing_if = "Option::is_none")]
-    pub idle_scaling: Option<bool>,
-    #[serde(rename = "idleTimeoutMinutes", skip_serializing_if = "Option::is_none")]
-    pub idle_timeout_minutes: Option<f64>,
-    #[serde(rename = "ipAccessList", skip_serializing_if = "Option::is_none")]
-    pub ip_access_list: Option<Vec<IpAccessListEntryResponse>>,
-    #[serde(rename = "isPrimary", skip_serializing_if = "Option::is_none")]
-    pub is_primary: Option<bool>,
-    #[serde(rename = "isReadonly", skip_serializing_if = "Option::is_none")]
-    pub is_readonly: Option<bool>,
-    #[serde(rename = "maxReplicaMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub max_replica_memory_gb: Option<f64>,
-    #[serde(rename = "maxReplicas", skip_serializing_if = "Option::is_none")]
-    pub max_replicas: Option<i64>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(rename = "maxTotalMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub max_total_memory_gb: Option<f64>,
-    #[serde(rename = "minReplicaMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub min_replica_memory_gb: Option<f64>,
-    #[serde(rename = "minReplicas", skip_serializing_if = "Option::is_none")]
-    pub min_replicas: Option<i64>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(rename = "minTotalMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub min_total_memory_gb: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(rename = "numReplicas", skip_serializing_if = "Option::is_none")]
-    pub num_replicas: Option<i64>,
-    #[serde(rename = "privateEndpointIds", skip_serializing_if = "Option::is_none")]
-    pub private_endpoint_ids: Option<Vec<String>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub profile: Option<ServiceProfile>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub provider: Option<ServiceProvider>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub region: Option<ServiceRegion>,
-    #[serde(rename = "releaseChannel", skip_serializing_if = "Option::is_none")]
-    pub release_channel: Option<ServiceReleasechannel>,
-    #[serde(rename = "replicaMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub replica_memory_gb: Option<f64>,
-    #[serde(rename = "scalingSchedule", skip_serializing_if = "Option::is_none")]
-    pub scaling_schedule: Option<ScalingSchedule>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub state: Option<ServiceState>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tags: Option<Vec<ResourceTagsV1Response>>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tier: Option<ServiceTier>,
-    #[serde(
-        rename = "transparentDataEncryptionKeyId",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub transparent_data_encryption_key_id: Option<String>,
-}
-
 /// `ServiceAccount` from the ClickHouse Cloud API.
 #[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
 pub struct ServiceAccount {
     #[serde(rename = "serviceAccountFile")]
     pub service_account_file: String,
-}
-
-/// `ServiceClickhouseSetting` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServiceClickhouseSetting {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub value: Option<String>,
-}
-
-/// `ServiceClickhouseSettingSchemaEntry` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServiceClickhouseSettingSchemaEntry {
-    #[serde(rename = "deprecationNotice", skip_serializing_if = "Option::is_none")]
-    pub deprecation_notice: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub description: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub r#enum: Option<Vec<i64>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub example: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub r#type: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub warning: Option<String>,
-}
-
-/// `ServiceClickhouseSettingWarning` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServiceClickhouseSettingWarning {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub message: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-}
-
-/// `ServiceClickhouseSettingsList` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServiceClickhouseSettingsList {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub settings: Option<Vec<ServiceClickhouseSetting>>,
-}
-
-/// `ServiceClickhouseSettingsPatchRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServiceClickhouseSettingsPatchRequest {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub settings: Option<String>,
-}
-
-/// `ServiceClickhouseSettingsPatchResponse` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServiceClickhouseSettingsPatchResponse {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub settings: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub warnings: Option<Vec<ServiceClickhouseSettingWarning>>,
-}
-
-/// `ServiceClickhouseSettingsSchema` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServiceClickhouseSettingsSchema {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub settings: Option<Vec<ServiceClickhouseSettingSchemaEntry>>,
-}
-
-/// `ServiceEndpoint` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-#[serde(from = "crate::serde_helpers::ServiceEndpointWire")]
-pub struct ServiceEndpoint {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub host: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    // The schema currently says `number`, but Cloud endpoints are TCP ports and
-    // the API sends integral values. Accept integral float syntax from the API,
-    // but keep JSON output integral for consumers.
-    pub port: Option<i64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub protocol: Option<ServiceEndpointProtocol>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub username: Option<String>,
-}
-
-/// `ServiceEndpointChange` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServiceEndpointChange {
-    pub enabled: bool,
-    pub protocol: ServiceEndpointChangeProtocol,
-}
-
-/// `ServicePasswordPatchRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServicePasswordPatchRequest {
-    #[serde(rename = "newDoubleSha1Hash", skip_serializing_if = "Option::is_none")]
-    pub new_double_sha1_hash: Option<String>,
-    #[serde(rename = "newPasswordHash", skip_serializing_if = "Option::is_none")]
-    pub new_password_hash: Option<String>,
-}
-
-/// `ServicePasswordPatchResponse` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServicePasswordPatchResponse {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub password: Option<String>,
-}
-
-/// `ServicePatchRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServicePatchRequest {
-    #[serde(rename = "enableCoreDumps", skip_serializing_if = "Option::is_none")]
-    pub enable_core_dumps: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub endpoints: Option<Vec<ServiceEndpointChange>>,
-    #[serde(rename = "ipAccessList", skip_serializing_if = "Option::is_none")]
-    pub ip_access_list: Option<IpAccessListPatch>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(rename = "privateEndpointIds", skip_serializing_if = "Option::is_none")]
-    pub private_endpoint_ids: Option<InstancePrivateEndpointsPatch>,
-    #[serde(rename = "releaseChannel", skip_serializing_if = "Option::is_none")]
-    pub release_channel: Option<ServicePatchRequestReleasechannel>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tags: Option<InstanceTagsPatch>,
-    #[serde(
-        rename = "transparentDataEncryptionKeyId",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub transparent_data_encryption_key_id: Option<String>,
-}
-
-/// `ServicePostRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServicePostRequest {
-    #[serde(rename = "autoscalingMode", skip_serializing_if = "Option::is_none")]
-    pub autoscaling_mode: Option<AutoscalingMode>,
-    #[serde(rename = "backupId", skip_serializing_if = "Option::is_none")]
-    pub backup_id: Option<uuid::Uuid>,
-    #[serde(rename = "byocId", skip_serializing_if = "Option::is_none")]
-    pub byoc_id: Option<String>,
-    #[serde(rename = "complianceType", skip_serializing_if = "Option::is_none")]
-    pub compliance_type: Option<ServicePostRequestCompliancetype>,
-    #[serde(rename = "dataWarehouseId", skip_serializing_if = "Option::is_none")]
-    pub data_warehouse_id: Option<String>,
-    #[serde(rename = "enableCoreDumps", skip_serializing_if = "Option::is_none")]
-    pub enable_core_dumps: Option<bool>,
-    #[serde(
-        rename = "encryptionAssumedRoleIdentifier",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub encryption_assumed_role_identifier: Option<String>,
-    #[serde(rename = "encryptionKey", skip_serializing_if = "Option::is_none")]
-    pub encryption_key: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub endpoints: Option<Vec<ServiceEndpointChange>>,
-    #[serde(
-        rename = "hasTransparentDataEncryption",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub has_transparent_data_encryption: Option<bool>,
-    #[serde(rename = "idleScaling", skip_serializing_if = "Option::is_none")]
-    pub idle_scaling: Option<bool>,
-    #[serde(rename = "idleTimeoutMinutes", skip_serializing_if = "Option::is_none")]
-    pub idle_timeout_minutes: Option<f64>,
-    #[serde(rename = "ipAccessList")]
-    pub ip_access_list: Vec<IpAccessListEntry>,
-    #[serde(rename = "isReadonly", skip_serializing_if = "Option::is_none")]
-    pub is_readonly: Option<bool>,
-    #[serde(rename = "maxReplicaMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub max_replica_memory_gb: Option<f64>,
-    #[serde(rename = "maxReplicas", skip_serializing_if = "Option::is_none")]
-    pub max_replicas: Option<i64>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(rename = "maxTotalMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub max_total_memory_gb: Option<f64>,
-    #[serde(rename = "minReplicaMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub min_replica_memory_gb: Option<f64>,
-    #[serde(rename = "minReplicas", skip_serializing_if = "Option::is_none")]
-    pub min_replicas: Option<i64>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(rename = "minTotalMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub min_total_memory_gb: Option<f64>,
-    pub name: String,
-    #[serde(rename = "numReplicas", skip_serializing_if = "Option::is_none")]
-    pub num_replicas: Option<i64>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(rename = "privateEndpointIds", skip_serializing_if = "Option::is_none")]
-    pub private_endpoint_ids: Option<Vec<String>>,
-    #[serde(
-        rename = "privatePreviewTermsChecked",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub private_preview_terms_checked: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub profile: Option<ServicePostRequestProfile>,
-    pub provider: ServicePostRequestProvider,
-    pub region: ServicePostRequestRegion,
-    #[serde(rename = "releaseChannel", skip_serializing_if = "Option::is_none")]
-    pub release_channel: Option<ServicePostRequestReleasechannel>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tags: Option<Vec<ResourceTagsV1>>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tier: Option<ServicePostRequestTier>,
-}
-
-/// `ServicePostResponse` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServicePostResponse {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub password: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub service: Option<Service>,
-}
-
-/// `ServiceQueryAPIEndpoint` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServiceQueryAPIEndpoint {
-    #[serde(rename = "allowedOrigins", skip_serializing_if = "Option::is_none")]
-    pub allowed_origins: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<String>,
-    #[serde(rename = "openApiKeys", skip_serializing_if = "Option::is_none")]
-    pub open_api_keys: Option<Vec<String>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub roles: Option<Vec<String>>,
-}
-
-/// `ServiceReplicaScalingPatchRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServiceReplicaScalingPatchRequest {
-    #[serde(rename = "autoscalingMode", skip_serializing_if = "Option::is_none")]
-    pub autoscaling_mode: Option<AutoscalingMode>,
-    #[serde(rename = "idleScaling", skip_serializing_if = "Option::is_none")]
-    pub idle_scaling: Option<bool>,
-    #[serde(rename = "idleTimeoutMinutes", skip_serializing_if = "Option::is_none")]
-    pub idle_timeout_minutes: Option<f64>,
-    #[serde(rename = "maxReplicaMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub max_replica_memory_gb: Option<f64>,
-    #[serde(rename = "maxReplicas", skip_serializing_if = "Option::is_none")]
-    pub max_replicas: Option<i64>,
-    #[serde(rename = "minReplicaMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub min_replica_memory_gb: Option<f64>,
-    #[serde(rename = "minReplicas", skip_serializing_if = "Option::is_none")]
-    pub min_replicas: Option<i64>,
-    #[serde(rename = "numReplicas", skip_serializing_if = "Option::is_none")]
-    pub num_replicas: Option<i64>,
-}
-
-/// `ServiceScalingPatchRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServiceScalingPatchRequest {
-    #[serde(rename = "idleScaling", skip_serializing_if = "Option::is_none")]
-    pub idle_scaling: Option<bool>,
-    #[serde(rename = "idleTimeoutMinutes", skip_serializing_if = "Option::is_none")]
-    pub idle_timeout_minutes: Option<f64>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(rename = "maxTotalMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub max_total_memory_gb: Option<f64>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(rename = "minTotalMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub min_total_memory_gb: Option<f64>,
-    #[serde(rename = "numReplicas", skip_serializing_if = "Option::is_none")]
-    pub num_replicas: Option<i64>,
-}
-
-/// `ServiceScalingPatchResponse` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServiceScalingPatchResponse {
-    #[serde(
-        rename = "availablePrivateEndpointIds",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub available_private_endpoint_ids: Option<Vec<String>>,
-    #[serde(rename = "autoscalingMode", skip_serializing_if = "Option::is_none")]
-    pub autoscaling_mode: Option<AutoscalingMode>,
-    #[serde(rename = "byocId", skip_serializing_if = "Option::is_none")]
-    pub byoc_id: Option<String>,
-    #[serde(rename = "clickhouseVersion", skip_serializing_if = "Option::is_none")]
-    pub clickhouse_version: Option<String>,
-    #[serde(rename = "complianceType", skip_serializing_if = "Option::is_none")]
-    pub compliance_type: Option<ServiceScalingPatchResponseCompliancetype>,
-    #[serde(rename = "createdAt", skip_serializing_if = "Option::is_none")]
-    pub created_at: Option<chrono::DateTime<chrono::Utc>>,
-    #[serde(rename = "currentScaling", skip_serializing_if = "Option::is_none")]
-    pub current_scaling: Option<CurrentScaling>,
-    #[serde(rename = "dataWarehouseId", skip_serializing_if = "Option::is_none")]
-    pub data_warehouse_id: Option<String>,
-    #[serde(rename = "enableCoreDumps", skip_serializing_if = "Option::is_none")]
-    pub enable_core_dumps: Option<bool>,
-    #[serde(
-        rename = "encryptionAssumedRoleIdentifier",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub encryption_assumed_role_identifier: Option<String>,
-    #[serde(rename = "encryptionKey", skip_serializing_if = "Option::is_none")]
-    pub encryption_key: Option<String>,
-    #[serde(rename = "encryptionRoleId", skip_serializing_if = "Option::is_none")]
-    pub encryption_role_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub endpoints: Option<Vec<ServiceEndpoint>>,
-    #[serde(
-        rename = "hasTransparentDataEncryption",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub has_transparent_data_encryption: Option<bool>,
-    #[serde(rename = "iamRole", skip_serializing_if = "Option::is_none")]
-    pub iam_role: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub id: Option<uuid::Uuid>,
-    #[serde(rename = "idleScaling", skip_serializing_if = "Option::is_none")]
-    pub idle_scaling: Option<bool>,
-    #[serde(rename = "idleTimeoutMinutes", skip_serializing_if = "Option::is_none")]
-    pub idle_timeout_minutes: Option<f64>,
-    #[serde(rename = "ipAccessList", skip_serializing_if = "Option::is_none")]
-    pub ip_access_list: Option<Vec<IpAccessListEntryResponse>>,
-    #[serde(rename = "isPrimary", skip_serializing_if = "Option::is_none")]
-    pub is_primary: Option<bool>,
-    #[serde(rename = "isReadonly", skip_serializing_if = "Option::is_none")]
-    pub is_readonly: Option<bool>,
-    #[serde(rename = "maxReplicaMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub max_replica_memory_gb: Option<f64>,
-    #[serde(rename = "maxReplicas", skip_serializing_if = "Option::is_none")]
-    pub max_replicas: Option<i64>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(rename = "maxTotalMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub max_total_memory_gb: Option<f64>,
-    #[serde(rename = "minReplicaMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub min_replica_memory_gb: Option<f64>,
-    #[serde(rename = "minReplicas", skip_serializing_if = "Option::is_none")]
-    pub min_replicas: Option<i64>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(rename = "minTotalMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub min_total_memory_gb: Option<f64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub name: Option<String>,
-    #[serde(rename = "numReplicas", skip_serializing_if = "Option::is_none")]
-    pub num_replicas: Option<i64>,
-    #[serde(rename = "privateEndpointIds", skip_serializing_if = "Option::is_none")]
-    pub private_endpoint_ids: Option<Vec<String>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub profile: Option<ServiceScalingPatchResponseProfile>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub provider: Option<ServiceScalingPatchResponseProvider>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub region: Option<ServiceScalingPatchResponseRegion>,
-    #[serde(rename = "releaseChannel", skip_serializing_if = "Option::is_none")]
-    pub release_channel: Option<ServiceScalingPatchResponseReleasechannel>,
-    #[serde(rename = "replicaMemoryGb", skip_serializing_if = "Option::is_none")]
-    pub replica_memory_gb: Option<f64>,
-    #[serde(rename = "scalingSchedule", skip_serializing_if = "Option::is_none")]
-    pub scaling_schedule: Option<ScalingSchedule>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub state: Option<ServiceScalingPatchResponseState>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tags: Option<Vec<ResourceTagsV1Response>>,
-    #[cfg(feature = "deprecated-fields")]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tier: Option<ServiceScalingPatchResponseTier>,
-    #[serde(
-        rename = "transparentDataEncryptionKeyId",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub transparent_data_encryption_key_id: Option<String>,
-}
-
-/// `ServiceStatePatchRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct ServiceStatePatchRequest {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub command: Option<ServiceStatePatchRequestCommand>,
-}
-
-/// `UpgradeWindow` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct UpgradeWindow {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub duration: Option<i64>,
-    #[serde(rename = "startHourUtc", skip_serializing_if = "Option::is_none")]
-    pub start_hour_utc: Option<i64>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub weekday: Option<i64>,
-}
-
-/// `UpgradeWindowPutRequest` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct UpgradeWindowPutRequest {
-    #[serde(rename = "startHourUtc")]
-    pub start_hour_utc: i64,
-    pub weekday: i64,
 }
 
 /// `UpdateReversePrivateEndpoint` from the ClickHouse Cloud API.
@@ -14138,295 +10736,6 @@ pub struct UpdateReversePrivateEndpoint {
         skip_serializing_if = "Option::is_none"
     )]
     pub custom_private_dns_mappings: Option<Vec<CustomPrivateDnsMapping>>,
-}
-
-/// `UsageCost` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct UsageCost {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub costs: Option<Vec<UsageCostRecord>>,
-    #[serde(rename = "grandTotalCHC", skip_serializing_if = "Option::is_none")]
-    pub grand_total_chc: Option<f64>,
-}
-
-/// `UsageCostMetrics` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct UsageCostMetrics {
-    #[serde(rename = "backupCHC", skip_serializing_if = "Option::is_none")]
-    pub backup_chc: Option<f64>,
-    #[serde(rename = "computeCHC", skip_serializing_if = "Option::is_none")]
-    pub compute_chc: Option<f64>,
-    #[serde(rename = "dataTransferCHC", skip_serializing_if = "Option::is_none")]
-    pub data_transfer_chc: Option<f64>,
-    #[serde(rename = "initialLoadCHC", skip_serializing_if = "Option::is_none")]
-    pub initial_load_chc: Option<f64>,
-    #[serde(
-        rename = "interRegionTier1DataTransferCHC",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub inter_region_tier1_data_transfer_chc: Option<f64>,
-    #[serde(
-        rename = "interRegionTier2DataTransferCHC",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub inter_region_tier2_data_transfer_chc: Option<f64>,
-    #[serde(
-        rename = "interRegionTier3DataTransferCHC",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub inter_region_tier3_data_transfer_chc: Option<f64>,
-    #[serde(
-        rename = "interRegionTier4DataTransferCHC",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub inter_region_tier4_data_transfer_chc: Option<f64>,
-    #[serde(
-        rename = "publicDataTransferCHC",
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub public_data_transfer_chc: Option<f64>,
-    #[serde(rename = "storageCHC", skip_serializing_if = "Option::is_none")]
-    pub storage_chc: Option<f64>,
-}
-
-/// `UsageCostRecord` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct UsageCostRecord {
-    #[serde(rename = "dataWarehouseId", skip_serializing_if = "Option::is_none")]
-    pub data_warehouse_id: Option<uuid::Uuid>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub date: Option<String>,
-    #[serde(rename = "entityId", skip_serializing_if = "Option::is_none")]
-    pub entity_id: Option<uuid::Uuid>,
-    #[serde(rename = "entityName", skip_serializing_if = "Option::is_none")]
-    pub entity_name: Option<String>,
-    #[serde(rename = "entityType", skip_serializing_if = "Option::is_none")]
-    pub entity_type: Option<UsageCostRecordEntitytype>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub locked: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub metrics: Option<UsageCostMetrics>,
-    #[serde(rename = "serviceId", skip_serializing_if = "Option::is_none")]
-    pub service_id: Option<uuid::Uuid>,
-    #[serde(rename = "totalCHC", skip_serializing_if = "Option::is_none")]
-    pub total_chc: Option<f64>,
-}
-
-/// `pgBouncerConfig` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PgBouncerConfig {}
-
-/// `pgBouncerConfig` from the ClickHouse Cloud API, in response position.
-///
-/// Response variant of [`PgBouncerConfig`]: every field is `Option<T>`, so a
-/// field the API drops or sends as `null` deserializes to `None` instead of
-/// failing. The schema currently declares no properties.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PgBouncerConfigResponse {}
-
-/// `pgConfig` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PgConfig {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub autovacuum_analyze_scale_factor: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub autovacuum_max_workers: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub autovacuum_naptime: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub autovacuum_vacuum_cost_delay: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub autovacuum_vacuum_cost_limit: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub autovacuum_vacuum_insert_scale_factor: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub autovacuum_vacuum_scale_factor: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub autovacuum_work_mem: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub default_transaction_isolation: Option<PgConfigDefaultTransactionIsolation>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub effective_cache_size: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub effective_io_concurrency: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub idle_in_transaction_session_timeout: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub idle_session_timeout: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub lock_timeout: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub maintenance_work_mem: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_connections: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_parallel_maintenance_workers: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_parallel_workers: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_parallel_workers_per_gather: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_slot_wal_keep_size: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_wal_size: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_worker_processes: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub min_wal_size: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub random_page_cost: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ssl_min_protocol_version: Option<PgConfigSslMinProtocolVersion>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub statement_timeout: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub transaction_timeout: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub wal_compression: Option<PgConfigWalCompression>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub wal_keep_size: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub wal_sender_timeout: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub work_mem: Option<serde_json::Value>,
-}
-
-/// `pgConfig` from the ClickHouse Cloud API, in response position.
-///
-/// Response variant of [`PgConfig`]: every field is `Option<T>`, so a field the
-/// API drops or sends as `null` deserializes to `None` instead of failing.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PgConfigResponse {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub autovacuum_analyze_scale_factor: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub autovacuum_max_workers: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub autovacuum_naptime: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub autovacuum_vacuum_cost_delay: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub autovacuum_vacuum_cost_limit: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub autovacuum_vacuum_insert_scale_factor: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub autovacuum_vacuum_scale_factor: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub autovacuum_work_mem: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub default_transaction_isolation: Option<PgConfigDefaultTransactionIsolation>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub effective_cache_size: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub effective_io_concurrency: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub idle_in_transaction_session_timeout: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub idle_session_timeout: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub lock_timeout: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub maintenance_work_mem: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_connections: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_parallel_maintenance_workers: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_parallel_workers: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_parallel_workers_per_gather: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_slot_wal_keep_size: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_wal_size: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_worker_processes: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub min_wal_size: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub random_page_cost: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub ssl_min_protocol_version: Option<PgConfigSslMinProtocolVersion>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub statement_timeout: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub transaction_timeout: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub wal_compression: Option<PgConfigWalCompression>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub wal_keep_size: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub wal_sender_timeout: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub work_mem: Option<serde_json::Value>,
-}
-
-/// `postgresInstanceConfig` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresInstanceConfig {
-    #[serde(rename = "pgBouncerConfig")]
-    pub pg_bouncer_config: PgBouncerConfig,
-    #[serde(rename = "pgConfig")]
-    pub pg_config: PgConfig,
-}
-
-/// `postgresInstanceConfig` from the ClickHouse Cloud API, in response
-/// position.
-///
-/// Response variant of [`PostgresInstanceConfig`]: every field is `Option<T>`,
-/// so a field the API drops or sends as `null` deserializes to `None` instead
-/// of failing. Writing a fetched configuration back to the API goes through
-/// `TryFrom<PostgresInstanceConfigResponse>` (see [`crate::convert`]), which
-/// forces every absent required field to be resolved explicitly.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresInstanceConfigResponse {
-    #[serde(rename = "pgBouncerConfig", skip_serializing_if = "Option::is_none")]
-    pub pg_bouncer_config: Option<PgBouncerConfigResponse>,
-    #[serde(rename = "pgConfig", skip_serializing_if = "Option::is_none")]
-    pub pg_config: Option<PgConfigResponse>,
-}
-
-/// `postgresInstanceUpdateConfigResponse` from the ClickHouse Cloud API.
-#[derive(Debug, Clone, PartialEq, Default, Serialize, Deserialize)]
-pub struct PostgresInstanceUpdateConfigResponse {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub message: Option<String>,
-    #[serde(rename = "pgBouncerConfig", skip_serializing_if = "Option::is_none")]
-    pub pg_bouncer_config: Option<PgBouncerConfigResponse>,
-    #[serde(rename = "pgConfig", skip_serializing_if = "Option::is_none")]
-    pub pg_config: Option<PgConfigResponse>,
-}
-
-impl Default for BackupBucket {
-    fn default() -> Self {
-        // Every field of a response variant is `Option<T>`, so the derived
-        // `AwsBackupBucket::default()` leaves `bucketProvider` absent and
-        // serializes to `{}` — which deserializes back through the
-        // discriminator dispatch as `Unknown`, not as this variant. Naming the
-        // variant's own wire value keeps the default round-tripping.
-        Self::AwsBackupBucket(AwsBackupBucket {
-            bucket_provider: Some(AwsBackupBucketBucketprovider::default()),
-            ..AwsBackupBucket::default()
-        })
-    }
-}
-
-impl Default for BackupBucketPatchRequest {
-    fn default() -> Self {
-        Self::AwsBackupBucketPatchRequestV1(AwsBackupBucketPatchRequestV1::default())
-    }
-}
-
-impl Default for BackupBucketPostRequest {
-    fn default() -> Self {
-        Self::AwsBackupBucketPostRequestV1(AwsBackupBucketPostRequestV1::default())
-    }
-}
-
-impl Default for BackupBucketProperties {
-    fn default() -> Self {
-        Self::AwsBackupBucketProperties(AwsBackupBucketProperties::default())
-    }
 }
 
 impl Default for ClickStackAlertChannel {
